### 2024-07-26 #7045 @editorialbot
[PRE REVIEW]: NcCut: A NetCDF Viewer and Transecting Tool
**Submitting author:** @rchartra (<a href="http://orcid.org/0000-0001-6948-8380">Robin Chartrand</a>)
**Repository:** https://github.com/rchartra/NcCut
**Branch with paper.md** (empty if default branch):
**Version:** v0.4.3
