### 2024-07-29 053cdd9
Coords for transect points come from NetCDF coords
