+++ tests/test_functions.py
@@ -24,7 +24,7 @@ class Test(unittest.TestCase):
         points = [1000, 200, 1200, 200]
 
         # App result
-        app = func.ip_get_points(points, img, False)["Cut"]
+        app = func.ip_get_points(points, img, {"image": SUPPORT_FILE_PATH + "example.jpg"})["Cut"]
 
         # Manual result
         arr = np.asarray(img)
@@ -42,7 +42,7 @@ class Test(unittest.TestCase):
         points = [1000, 200, 1200, 400]
 
         # App result
-        app = func.ip_get_points(points, img, False)["Cut"]
+        app = func.ip_get_points(points, img, {"image": SUPPORT_FILE_PATH + "example.jpg"})["Cut"]
 
         # Manual result
         arr = np.asarray(img)
@@ -63,7 +63,7 @@ class Test(unittest.TestCase):
         points = [1000, 100, 1000, 400]
 
         # App result
-        app = func.ip_get_points(points, img, False)["Cut"]
+        app = func.ip_get_points(points, img, {"image": SUPPORT_FILE_PATH + "example.jpg"})["Cut"]
 
         # Manual result
         arr = np.asarray(img)
@@ -78,56 +78,74 @@ class Test(unittest.TestCase):
         Test an accurate transect is made when taken horizontally on a NetCDF file
         """
         # Setup
-        dat = xr.open_dataset(SUPPORT_FILE_PATH + "example_4v.nc")['Divergence'].data
-        points = [1000, 200, 1200, 200]
+        dat = xr.open_dataset(SUPPORT_FILE_PATH + "example_3d.nc")['Theta'].sel(k=0)
+        config = {"netcdf":{"x": "i", "y": "j", "z": "k", "z_val": "0", "var": "Theta", "file": dat}}
+        points = [100, 50, 200, 50]
 
         # App result
-        app = func.ip_get_points(points, dat, True)["Cut"]
+        app = func.ip_get_points(points, dat.data, config)
 
         # Manual result
-        arr = np.asarray(dat)
+        arr = np.asarray(dat.data)
         rows = np.shape(arr)[0]
         manual = arr[rows - points[3] - 1:rows - points[1], points[0]:points[2]][0]
         # Compare
-        self.assertEqual(max(app - manual), 0, "Transect on NetCDF not accurate at zero degrees")
+        self.assertEqual(max(app["Cut"] - manual), 0, "Transect on NetCDF not accurate at zero degrees")
+        # Check Coordinates from NetCDF
+        self.assertListEqual(list(dat.coords["i"][points[0]:points[2]]), app["i"],
+                             "X Coordinates for NetCDF 0 Degree Transect Incorrect")
+        self.assertListEqual(np.repeat(dat.coords["j"][points[1]].data, len(manual)).tolist(), app["j"],
+                             "Y Coordinates for NetCDF 0 DegreeTransect Incorrect")
 
     def test_transect_45_deg_nc(self):
         """
         Test an accurate transect is made when taken at 45 degrees on a NetCDF file
         """
         # Setup
-        dat = xr.open_dataset(SUPPORT_FILE_PATH + "example_4v.nc")['Divergence'].data
-        points = [1000, 200, 1200, 400]
+        dat = xr.open_dataset(SUPPORT_FILE_PATH + "example_3d.nc")['Theta'].sel(k=0)
+        config = {"netcdf":{"x": "i", "y": "j", "z": "k", "z_val": "0", "var": "Theta", "file": dat}}
+        points = [100, 50, 200, 150]
 
         # App result
-        app = func.ip_get_points(points, dat, True)["Cut"]
+        app = func.ip_get_points(points, dat.data, config)
 
[…]
