### 2024-07-29 1765705
Coords for transect points come from NetCDF coords
