### 2024-07-29 090eba5
Coords for transect points come from NetCDF coords
