### 2024-07-29 532ee84
Double Click to start new marker. Can now click on labels for options in plotting menu to select them.
