### 2024-07-29 cd94fa8
Upgrade touch mimicking class
