### 2024-07-30 #7045 @hugoledoux
@kthyng @rchartra I'm going at the end of this busy week 10 days on holidays, I'll try to start this before Friday but I doubt it'll happen.

In the meantime, perhaps you could provide 5 potential reviewers from our database (https://reviewers.joss.theoj.org/lookup) or outside, put their github usernames without the '@' so that they don't get pinged needlessly.
