### 2024-08-15 0614ca7
Change default circle size to 5
