### 2024-08-15 6ee9e68
Improved number positioning when graphic size is changed
