+++ src/nccut/plotpopup.py
@@ -66,8 +66,8 @@ class PlotPopup(Popup):
     Attributes:
         home: Reference to root :class:`nccut.homescreen.HomeScreen` instance.
         all_transects (dict): Dictionary of data from all transects marked out by the user.
-        t_type (str): 'Marker' if transects came from transect marker tool or 'Multi' if transects
-            came from transect tool.
+        t_type (str): 'Marker' if transects came from transect marker tool, 'Chain' if transects came from transect
+            chain tool or 'Multi' if transects came from transect tool.
         active_transects (dict): Dictionary of currently selected transects. 'Click <X Cord>', 'Click <Y Cord>', and
             'Width' fields should be removed (if marker tool) to simplify plotting. Contains average
             of transects if marker tool was used with a constant transect width.
@@ -105,6 +105,8 @@ class PlotPopup(Popup):
         self.config = config
         if list(self.all_transects.keys())[0][0:-2] == "Marker":
             self.t_type = "Marker"
+        elif list(self.all_transects.keys())[0][0:-2] == "Chain":
+            self.t_type = "Chain"
         else:
             self.t_type = "Multi"
 
@@ -114,8 +116,12 @@ class PlotPopup(Popup):
         act = copy.deepcopy(self.all_transects)
         for key in list(act.keys()):
             self.active_transects[key] = {}
-            if self.t_type == "Marker":  # Remove fields that don't get plotted
+            # Remove fields that don't get plotted
+            if self.t_type == "Marker":
                 for k in list(act[key].keys())[:3]:
+                    act[key].pop(k)
+            elif self.t_type == "Chain":
+                for k in list(act[key].keys())[:2]:
                     act[key].pop(k)
             self.active_transects[key] = dict.fromkeys(act[key], False)
 
@@ -127,11 +133,11 @@ class PlotPopup(Popup):
                     new.update(self.active_transects[key])
                     self.active_transects[key] = new
 
-        # If not marker, start will all transects selected
+        # Start with the first group selected
         first = list(self.active_transects.keys())[0]
         self.active_transects[first] = dict.fromkeys(self.active_transects[first], True)
 
-        # If marker start with first marker selected
+        # If marker start with average not selected
         if self.t_type == "Marker":
             w_lis = self.all_transects[first]['Width']
             if all(x == w_lis[0] for x in w_lis):
@@ -200,8 +206,8 @@ class PlotPopup(Popup):
         t_box.add_widget(t_lab)
         self.t_select = func.RoundedButton(text="Select...", size_hint=(0.5, 1), font_size=self.home.font)
         t_box.add_widget(self.t_select)
-        if self.t_type == "Marker":
-            t_drop = self.get_marker_dropdown()
+        if self.t_type == "Marker" or self.t_type == "Chain":
+            t_drop = self.get_group_dropdown()
         else:
             t_drop = self.get_cut_dropdown('Multi')
         self.t_select.bind(on_press=lambda x: t_drop.open(self.t_select))
@@ -371,17 +377,17 @@ class PlotPopup(Popup):
         else:  # Build correctly formatted JSON file
             dat = copy.copy(self.active_data)
             if len(self.active_vars) == 0:  # If Image
-                final = self.add_marker_info(dat)
+                final = self.add_group_info(dat)
             elif len(self.active_z) == 0:  # If 2D NetCDF
                 final = {}
                 for var in list(dat.keys()):
-                    final[var] = self.add_marker_info(dat[var])
+                    final[var] = self.add_group_info(dat[var])
             else:  # If 3D NetCDF
                 final = {}
                 for var in list(dat.keys()):
                     final[var] = {}
                     for z in list(dat[var].keys()):
-                        final[var][z] = self.add_marker_info(dat[var][z])
[…]
