### 2024-08-19 0e2c5a0
Added new tool that allows multiple transects to be chained together
