+++ src/nccut/multichain.py
@@ -1,0 +1,205 @@
+"""
+Transect chain tool widget.
+
+Manages having multiple chains on screen at once.
+"""
+
+import kivy.uix as ui
+from kivy.core.window import Window
+from scipy.interpolate import CubicSpline
+import nccut.functions as func
+from nccut.plotpopup import PlotPopup
+from nccut.chain import Chain
+
+
+class MultiChain(ui.widget.Widget):
+    """
+    Transect chain tool widget.
+
+    Created when Transect Chain button is selected. From there on this object manages the creation,
+    modification, and data packaging of chains.
+
+    Attributes:
+        c_on (bool): Whether there are any chains active
+        home: Reference to root :class:`nccut.homescreen.HomeScreen` instance
+        dbtn: RoundedButton, Plot button to activate PlotPopup
+        dragging (bool): Whether viewer is in dragging mode
+        clicks (int): Number of clicks made by user. Does not decrease when points are deleted
+            unless all points are deleted in which case it goes back to zero.
+        nbtn: RoundedButton, New chain button
+        plotting: :class:`nccut.plotpopup.PlotPopup`, reference to plotting menu when opened
+    """
+    def __init__(self, home, **kwargs):
+        """
+        Defines sidebar elements and initializes widget
+
+        Args:
+            home: Reference to root :class:`nccut.homescreen.HomeScreen` instance
+        """
+        super(MultiChain, self).__init__(**kwargs)
+        self.c_on = False
+        self.home = home
+        self.dbtn = func.RoundedButton(text="Plot", size_hint=(1, 0.1), font_size=self.home.font)
+        self.dbtn.bind(on_press=lambda x: self.gather_popup())
+        self.dragging = False
+        self.clicks = 0
+        self.plotting = None
+
+        # New Chain Button
+        self.nbtn = func.RoundedButton(text="New Chain", size_hint=(1, 0.1), font_size=self.home.font)
+
+        self.nbtn.bind(on_press=lambda x: self.new_chain())
+        self.home.ids.sidebar.add_widget(self.nbtn, 1)
+
+    def font_adapt(self, font):
+        """
+        Updates font of sidebar elements and plotting menu if loaded.
+        Args:
+            font (float): New font size
+        """
+        self.dbtn.font_size = font
+        self.nbtn.font_size = font
+        if self.plotting:
+            self.plotting.font_adapt(font)
+
+    def update_l_col(self, color):
+        """
+        Asks each chain to update their line color
+
+        Args:
+            color (str): New color value: 'Blue', 'Green' or 'Orange'
+        """
+        for m in self.children:
+            m.update_l_col(color)
+
+    def update_c_size(self, value):
+        """
+       Asks each chain to update their circle size
+
+       Args:
+           value (float): New circle size
+       """
+        for m in self.children:
+            m.update_c_size(value)
+
+    def change_dragging(self, val):
+        """
+        Change whether in dragging mode or not.
+
+        Args:
+            val (bool): Whether in dragging mode or not.
+        """
+        self.dragging = val
+
+    def del_line(self):
+        """
+        Delete most recent chain with some safeguards.
+
+        If only one chain on screen, delete but add new chain and remove sidebar elements.
+        Otherwise delete current chain and go to previous. If no chains are on screen nothing
+        happens.
+        """
+        if len(self.children) == 0:
+            # If no chains on screen do nothing
+            return
+        Window.unbind(mouse_pos=self.children[0].draw_line)
+        self.remove_widget(self.children[0])
+        if len(self.children) == 0:
+            # Remove sidebar buttons if deleted chain was the only chain
+            self.clicks = 0
+            if self.dbtn in self.home.display.current:
[…]
