+++ src/nccut/chain.py
@@ -1,0 +1,200 @@
+"""
+Singular chain widget.
+
+Graphics and functionality of a singular chain created by the transect chain tool.
+"""
+
+import kivy.uix as ui
+from kivy.metrics import dp
+from kivy.graphics import Color, Ellipse, Line
+from kivy.uix.label import Label
+from kivy.core.window import Window
+from nccut.multitransect import MultiTransect
+from nccut.singletransect import SingleTransect
+
+
+class Chain(ui.widget.Widget):
+    """
+    Singular transect chain widget.
+
+    Graphics and functionality of a singular chain created by the transect chain tool. Draws point on each user click
+    and connects points with a line creating a chain of transects. Uses a MultiTransect object to store ana manage the
+    individual transects.
+
+    Attributes:
+        clicks (int): Number of clicks user has made. Decreases when points are deleted.
+        points (list): List of Tuples, For each click user makes: (X-coord, Y-coord, t_width).
+        home: Reference to root :class:`nccut.homescreen.HomeScreen` instance
+        base: :class:`nccut.multitransect.MultiTransect` object that manages transects
+        curr_line: kivy.graphics.Line, Line between cursor and last clicked
+        number: kivy.uix.label.Label, Reference to the number label
+        size: 2 element array of ints, Size of widget
+        pos: 2 element array of ints, Position of widget
+        l_color: kivy.graphics.Color, Color to use for graphics
+        c_size: 2 element tuple of floats that defines size of circles
+        line_width (float): Width of lines
+    """
+    def __init__(self, home, **kwargs):
+        """
+        Sets initial settings and initializes object.
+
+        Args:
+            home: Reference to root :class:`nccut.homescreen.HomeScreen` instance
+        """
+        super(Chain, self).__init__(**kwargs)
+        self.clicks = 0
+        self.points = []
+        self.home = home
+        self.base = MultiTransect(home=self.home)
+        self.curr_line = Line()
+        self.number = None
+        self.size = self.home.display.size
+        self.pos = self.home.display.pos
+        color = self.home.display.l_col
+        if color == "Blue":
+            self.l_color = Color(0.28, 0.62, 0.86)
+        elif color == "Green":
+            self.l_color = Color(0.39, 0.78, 0.47)
+        elif color == "Orange":
+            self.l_color = Color(0.74, 0.42, 0.13)
+        size = self.home.display.cir_size
+        self.c_size = (dp(size), dp(size))
+        self.line_width = dp(size / 5)
+
+    def update_l_col(self, color):
+        """
+        Update the line color and redraw all items on canvas.
+
+        Args:
+            color (str): New line color to use
+        """
+        if color == "Blue":
+            self.l_color = Color(0.28, 0.62, 0.86)
+        elif color == "Green":
+            self.l_color = Color(0.39, 0.78, 0.47)
+        elif color == "Orange":
+            self.l_color = Color(0.74, 0.42, 0.13)
+        groups = []
+        for c in range(1, self.clicks + 1):
+            group = self.canvas.get_group(str(c))
+            groups.append(group)
+        self.canvas.clear()
+        self.remove_widget(self.number)
+        self.canvas.add(self.l_color)
+        for g in groups:
+            for i in g:
+                self.canvas.add(i)
+        self.canvas.add(self.curr_line)
+        if self.clicks > 0:
+            self.add_widget(self.number)
+
+    def update_c_size(self, value):
+        """
+        Updates graphic sizes according to the new circle size value.
+
+        Args:
+            value (float): New graphics size
+        """
+        self.line_width = dp(value / 5)
+        for c in range(1, self.clicks + 1):
+            group = self.canvas.get_group(str(c))
+            for i in group:
+                if isinstance(i, Ellipse):
+                    i.size = (dp(value), dp(value))
[…]
