### 2024-08-20 #7045 @platipodium
@hugoledoux I'd be interested to review pending completion of my currently ongoing review work for JOSS https://github.com/openjournals/joss-reviews/issues/7025
