### 2024-08-27 85ea1fd
Better settings and performance improvements
