+++ src/nccut/functions.py
@@ -17,6 +17,7 @@ import numpy as np
 import math
 import re
 from pathlib import Path
+import xarray as xr
 
 
 class RoundedButton(Button):
@@ -126,21 +127,63 @@ def sel_data(config):
             :meth:`nccut.netcdfconfig.NetCDFConfig.check_inputs`
 
     Returns:
-        2D array of data from NetCDF file.
+        2D Dataset from NetCDF file according to selections transposed to (y, x).
     """
     # IMPORTANT: Numpy indexes (row, column) with (0, 0) at top left corner
     if config['z'] == 'N/A':
         # 2D NetCDF data
         ds = config['file'][config['var']].rename({config['y']: 'y', config['x']: 'x'})
-        ds = ds.transpose('y', 'x')
         data = ds.sel(y=ds['y'], x=ds['x'])
     else:
         # 3D NetCDF data
         ds = config['file'][config['var']].rename({config['y']: 'y', config['x']: 'x', config['z']: 'z'})
-        ds = ds.transpose('y', 'x', 'z')
         ds['z'] = ds['z'].astype(str)
         data = ds.sel(y=ds['y'], x=ds['x'], z=config['z_val'])
-    return np.flip(data.data, 0)
+    return data.transpose('y', 'x')
+
+
+def subset_around_transect(ds, points):
+    """
+    Determines a subset of the data with appropriate margins that surround the transect.
+
+    If room is available around
+
+    Args:
+        ds: Data to be subset from. Either a DataArray (X and Y coordinates must be named "x" and "y") or a 2D array
+        points: 4 element 1D array with coordinates of the two transect end points: [X1, Y1, X2, Y2]
+
+    Returns:
+        Subset of data, rescaled points [X1, Y1, X2, Y2], list of [Y, X] rescale factors
+    """
+    # Determine x, y bounds for subset area around transect
+    ys_xs = [sorted([points[1], points[3]]), sorted([points[0], points[2]])]
+    ys_xs = [c for cs in ys_xs for c in cs]
+    if isinstance(ds, xr.DataArray):
+        margins = [len(ds.coords["y"].data), len(ds.coords["x"].data)]
+    else:
+        margins = ds.shape
+    # Include a buffer margin if possible for interpolating
+    for i, p in enumerate(ys_xs):
+        if i % 2 == 0:  # Minimums
+            if p < 3:
+                ys_xs[i] = int(np.floor(p))
+            else:
+                ys_xs[i] = int(np.floor(p)) - 3
+        else:  # Maximums
+            if p > margins[int(i / 3)] - 3:
+                ys_xs[i] = int(np.floor(p))
+            else:
+                ys_xs[i] = int(np.floor(p)) + 3
+    # Subset dataset around the transect
+    if isinstance(ds, xr.DataArray):
+        ds = ds.isel({"y": slice(ys_xs[0], ys_xs[1] + 1), "x": slice(ys_xs[2], ys_xs[3] + 1)})
+    else:
+        ds = ds[(margins[0] - 1 - ys_xs[1]): (margins[0] - ys_xs[0]), ys_xs[2]: ys_xs[3] + 1]
+
+    points = [points[0] - ys_xs[2], points[1] - ys_xs[0], points[2] - ys_xs[2], points[3] - ys_xs[0]]
+
+    scales = [ys_xs[0], ys_xs[2]]
+    return ds, points, scales
 
 
 def ip_get_points(line, curr, config):
@@ -148,61 +191,34 @@ def ip_get_points(line, curr, config):
     Creates a data frame containing x, y, and value of points on transect line
 
     From end points finds points of line connecting them. Interpolates higher resolution version of
-    relevant section of data set, and then grabs data values at each line point.
+    given data set, and then grabs data values at each line point. If the file is a NetCDF file with valid
[…]
