### 2024-08-27 911cf5f
Better settings and performance improvements
