### 2024-09-06 #7185 @editorialbot
[REVIEW]: NcCut: A NetCDF Viewer and Transecting Tool
**Submitting author:** @rchartra (<a href="http://orcid.org/0000-0001-6948-8380">Robin Chartrand</a>)
**Repository:** https://github.com/rchartra/NcCut
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
