### 2024-09-06 rchartra/NcCut#9 @platipodium
Motivate GPL license
This is part of my review at https://github.com/openjournals/joss-reviews/issues/7185

@rchartra you're using the strong copyleft GPL-v3 license

 * are there constraints that force you to use GPL-v3?
 * is there a reason not to use a more permissive license?

If you insist on GPL v3, can it be relaxed to GPL-v3 or any later version?
