### 2024-09-06 rchartra/NcCut#10 @platipodium
GPL not a suitable choice for pictures, data files, and texts
This is part of my review at https://github.com/openjournals/joss-reviews/issues/7185

While GPL is a popular open source license for source code, it does not apply well to text, pictures and data, and you have a lot of those in your repository.

- The JOSS paper.md, for example, will be released under a Creative Commons license.
- Choose a license for texts
- Choose a license for pictures and data
- Consider releasing all your files that have no important IP in the public domain, such as CC0-1.0. (which is applicable to software, text + data).
