### 2024-09-06 rchartra/NcCut#11 @platipodium
Confirm that all of the code, data, and text is your sole IP
This is part of my review at https://github.com/openjournals/joss-reviews/issues/7185

Currently, no other sources for data, text, and images are acknowledged.  So I must assume, you coded, wrote, and created all of the artifacts in your repository yourself and by yourself.  Did you really reinvent all the wheels?

Please confirm that this is the case.  If not, please
* acknowledge your contributors
* cite your sources
* clarify licensing issues with your sources and contributors.
