### 2024-09-06 rchartra/NcCut#11 @rchartra
All code, text, videos and images here and in my docs are my own work and mine alone. Some of the example data and images I provide are not mine though they come from free access sources. I can add formal citations to them in an acknowledgements section.
