### 2024-09-06 rchartra/NcCut#11 @rchartra
The packages I used in my code are all open source and free access. Do these need citations as well?
