### 2024-09-06 rchartra/NcCut#10 @rchartra
I am under no restrictions regarding licenses. If I switch my license over to CC0-1.0 will that cover everything?
