### 2024-09-06 rchartra/NcCut#10 @platipodium
> I am under no restrictions regarding licenses. If I switch my license over to CC0-1.0 will that cover everything?

Yes it would.  But I personally would do this (CC0-1.0) only for

* data and text without much creative input from your side
* code fragments like __init__.py pyproject.toml .gititnore Makefile and similar that have no or negligible creative input from your side

For everything that is the result of your creative process, choose a different open source license.  Have a look at [choosealicense.com](https://choosealicense.com) and make a wise choice.

Since many of your files will now have a different license, you now cannot have only one license in your LICENSE.md file, but you may have to add some license management.  I can recommend https://reuse.software
