### 2024-09-06 rchartra/NcCut#9 @platipodium
Don't get me wrong, it is perfectly fine to choose GPL, especially if you want strong copyleft of your code.  See [choosealicense.com](https://choosealicense.com) for more information. Also, check what your peers in your community prefer.
