### 2024-09-06 rchartra/NcCut#11 @platipodium
> The packages I used in my code are all open source and free access. Do these need citations as well?

No.  Software that you do not distribute yourself does not need to be cited or acknowledged.
