### 2024-09-06 rchartra/NcCut#9 @platipodium
One special instance when you would be bound to using GPL:

If *any* of the packages you use are GPL-licensed, then you have to make your code GPL, too, *if* (and only if) you want to distribute binaries (that would contain this GPL-licensed code).
