### 2024-09-06 rchartra/NcCut#11 @platipodium
> All code, text, videos and images here and in my docs are my own work and mine alone.

Perfect 👍

> Some of the example data and images I provide are not mine though they come from free access sources.

Free access does not mean that you can use them legally.  For each of those data, you need to check the license and you need to state the license and the copyright holder, if you want to distribute that data as part of your (source code) package.

>  I can add formal citations to them in an acknowledgements section.

That is usually not required, except when stated in the license.  But you can - of course - be nice and acknowledge anyway.
