+++ src/nccut/nccut.py
@@ -1,3 +1,7 @@
+# SPDX-FileCopyrightText: 2024 University of Washington
+#
+# SPDX-License-Identifier: BSD-3-Clause
+
 """
 Builds app and sets initial window.
 
