+++ tests/__init__.py
@@ -1,0 +1,3 @@
+# SPDX-FileCopyrightText: 2024 University of Washington
+#
+# SPDX-License-Identifier: BSD-3-Clause
