+++ docs/build/html/_sources/usage.rst.txt
@@ -1,136 +1,0 @@
-Usage
-=====
-
-#. To install the app refer to the :doc:`installation` instructions.
-
-#. To run the app there are two options:
-
-    * From the command Line:
-
-        .. code-block:: console
-
-            nccut
-
-        You can also pass a NetCDF or image file directly into the app via the command line:
-
-        .. code-block:: console
-
-            nccut file_name
-
-    * From Python:
-
-        .. code-block:: python
-
-            from nccut.nccut import NcCut
-            NcCut().run()
-
-For video walkthroughs of NcCut's basic usage, refer to the :doc:`example` section.
-
-Loading a File
---------------
-
-**NcCut** supports .jpg, .jpeg, and .png image files and NetCDF .nc files. To load the file you must either know the absolute path to your file on your system or move the files into the same directory as your code. In this case you can use the relative file path.
-
-#. To load an image or NetCDF file type the absolute or relative file path to your file in the file entry text box
-
-    * If you are loading a NetCDF file a popup window will appear with additional selections:
-
-        * Select which variable from your file you would like to see.
-        * Confirm or select which dimensions you would like to use as the X and Y axes for your variable.
-        * If your variable has data in a third dimension select the Z dimension and an initial Z value to display. You will be able to toggle between z values later.
-        * At the moment NcCut can only load variables with 2 or 3 dimensions.
-
-#. Select **Go** to load the selected data
-
-#. You can scroll to zoom in and out of the image and can click and drag the image to move it around.
-
-#. From the **View** menu in the settings bar you can rotate or flip the image as well as change graphic settings for the tools.
-
-#. If you are loading a NetCDF file, from the **NetCDF** menu in the settings bar you can change which variable or z value you'd like to see as well as change the color map or contrast of the image.
-
-    * Changing the contrast and colormap only affects the displayed image. Transect data is taken from the dataset itself.
-
-Tools
------
-
-**NcCut** has two types of tools for making transects. The simpler :ref:`Transect <transect>` tool allows you to individually draw lines across features where transects will be taken when plotted:
-
-    .. image:: _media/transect.png
-
-The :ref:`Transect Marker <transect_marker>` tool allows you to draw a line along a feature, and transects will be automatically made orthogonal to the line you've drawn with a set width:
-
-    .. image:: _media/transect_marker.png
-
-.. _transect:
-
-Transect Tool
-^^^^^^^^^^^^^
-
-#. Hit the **Transect** button to enter transect mode
-#. Click two points you'd like to make a transect between
-
-    * Select **Drag Mode** from the actions sidebar to drag the image without selecting points, and select **Transect Mode** to go back.
-    * Select **Edit Mode** to delete either the last point clicked or the last transect drawn.
-#. Repeat for as many transects as you'd like
-#. When done select the **Plot** button and a popup will appear with a plot of all transects and downloading options.
-
-.. _transect_marker:
-
-Transect Marker
-^^^^^^^^^^^^^^^
-
-Using this tool multiple lines or *Markers* can be drawn onto a loaded file. These *Markers* can all be saved together as a *project* and reloaded into the viewer later. When a large file is being explored this allows you to mark out features over multiple sessions.
-
-#. Hit the **Transect Marker** button to enter transect mode.
-#. Click points along the feature you'd like to make transects across. Dots will appear on either side of the line drawn indicating the start and end points of the transects that will be made.
-
-    * Select **Drag Mode** from the actions sidebar to drag the image without selecting points, and select **Transect Mode** to go back.
-    * Select **Edit Mode** to delete either the last point clicked or the last marker drawn.
-    * To change the width of the transects being made you may enter the number of pixels into the **Width** text box. Select **Go** to change the width for all future transects.
-
[…]
