+++ docs/build/html/_sources/modules.rst.txt
@@ -1,7 +1,0 @@
-cutview
-=======
-
-.. toctree::
-   :maxdepth: 4
-
-   cutview
