+++ REUSE.toml
@@ -1,0 +1,19 @@
+version = 1
+
+[[annotations]]
+
+path = "**"
+SPDX-FileCopyrightText = "2024 University of Washington"
+SPDX-License-Identifier = "BSD-3-Clause"
+
+[[annotations]]
+
+path = "images/*"
+SPDX-FileCopyrightText = "2024 University of Washington"
+SPDX-License-Identifier = "CC0-1.0"
+
+[[annotations]]
+
+path = "support/*"
+SPDX-FileCopyrightText = "2024 University of Washington"
+SPDX-License-Identifier = "CC0-1.0"
