### 2024-09-09 69d9772
Remove img2pdf dependency
