+++ src/nccut/plotpopup.py
@@ -29,7 +29,6 @@ import os
 import pandas as pd
 import json
 import cv2
-import img2pdf
 
 
 class BackgroundDropDown(DropDown):
@@ -354,10 +353,8 @@ class PlotPopup(Popup):
             ipath = self.home.rel_path / (file + ".png")
             ppath = self.home.rel_path / (file + ".pdf")
             self.plot.texture.save(str(ipath.absolute()))
-            img = cv2.flip(cv2.imread(str(ipath.absolute())), 0)
-            cv2.imwrite(str(ipath.absolute()), img)
-            with open(str(ppath.absolute()), "wb") as f:
-                f.write(img2pdf.convert(str(ipath.absolute())))
+            img = im.fromarray(cv2.flip(cv2.imread(str(ipath.absolute())), 0))
+            img.save(ppath, "PDF")
             os.remove(ipath)
             func.alert("Download Complete", self.home)
 
