### 2024-09-09 1006d9a
Remove img2pdf dependency
