### 2024-09-10 a58148c
Add license fo ESA Sentinel-1 SAR image
