### 2024-09-10 rchartra/NcCut#11 @rchartra
Alright so following UW's licensing recommendations I've switched over to the BSD 3 Clause license following the Reuse Software guidelines. UW indeed holds the copyright rights so I've added copyright statements indicating that. As for my non code files, my example data comes from the family of MITgcm LLC simulations which are all listed as being in the Public Domain. One of my example images comes from the ESA and I've followed their protocol laid out by their CC-BY-SA-3.0-IGO license. All my other images I created I've released under CC0-1.0.
