### 2024-09-19 03d8c8b
Add interactive plotting for all z image. Update docs for transect chain tool
