+++ tests/test_gui.py
@@ -136,6 +136,9 @@ class Test(unittest.TestCase):
                 self.assertListEqual(og_img, run_app.home.display.children[0].children, "All Tools were not removed")
 
     def test_file_cleanup(self):
+        """
+        Tests all tools, file elements, and sidebar additions are reset when a new file is opened.
+        """
         og_side = copy.copy(run_app.home.ids.sidebar.children)
 
         # Open a file and use a tool
@@ -144,6 +147,7 @@ class Test(unittest.TestCase):
         load_2d_nc("Vorticity")
         self.assertEqual(len(run_app.home.color_bar_box.children), 1, "Colorbar was not Added")
         self.assertIsNotNone(run_app.home.color_bar_box.parent, "Colorbar box was not displayed")
+        self.assertIsNotNone(run_app.home.netcdf_info.parent, "NetCDF Info bar was not displayed")
         select_sidebar_button("Transect Marker")
         x = run_app.home.size[0]
         y = run_app.home.size[1]
@@ -161,6 +165,7 @@ class Test(unittest.TestCase):
         self.assertEqual(og_side, run_app.home.ids.sidebar.children, "Sidebar was not restored")
         self.assertEqual(len(run_app.home.color_bar_box.children), 0, "Colorbar was not removed")
         self.assertIsNone(run_app.home.color_bar_box.parent, "Colorbar box was not removed")
+        self.assertIsNone(run_app.home.netcdf_info.parent, "NetCDF info bar was not removed")
         self.assertEqual(len(run_app.home.display.children), 1, "Not all tools were removed from display")
 
     def test_project_upload(self):
@@ -232,7 +237,7 @@ class Test(unittest.TestCase):
 
         # Add new line
         select_sidebar_button("Back")
-        select_sidebar_button("New Line")
+        select_sidebar_button("New Marker")
         self.assertEqual(len(multi_mark_instance.children), 3, "New marker was not created")
 
         # Delete point
@@ -245,7 +250,7 @@ class Test(unittest.TestCase):
 
         # Test width adjustments
         select_sidebar_button("Back")
-        select_sidebar_button("New Line")
+        select_sidebar_button("New Marker")
 
         # Simulate marker press with width changes
         x = run_app.home.size[0]
@@ -348,69 +353,6 @@ class Test(unittest.TestCase):
         select_sidebar_button("Delete Last Line")
         self.assertEqual(len(tran_instance.children), 1, "When last chain is deleted, a new chain is added")
         self.assertEqual(tran_instance.children[0].points, [], "Transect Chain was not properly deleted")
-
-    def test_multi_transect(self):
-        """
-        Test regular transect tool exhibits expected behavior
-        """
-        run_app.home.ids.file_in.text = SUPPORT_FILE_PATH + "example.jpg"
-        run_app.home.go_btn()
-        self.assertEqual(run_app.home.file_on, True, "File was not loaded")
-
-        # Open Transect tool
-        select_sidebar_button("Transect")
-
-        x = run_app.home.size[0]
-        y = run_app.home.size[1]
-        incs = np.array([0.4, 0.45, 0.55, 0.6])
-        x_arr = incs * x
-        y_arr = incs * y
-
-        tran_instance = run_app.home.display.tool
-        # First Click
-        sidebar = run_app.home.ids.sidebar
-        tran_instance.on_touch_down(Click(float(x_arr[0]), float(y_arr[0])))
-        self.assertIsNone(next((but for but in sidebar.children if but.text == "Plot"), None),
-                          "There cannot be a Plot Button on First Click")
-        self.assertEqual(len(tran_instance.children), 1, "Transect Not Added")
-
-        # Second Click
-        tran_instance.on_touch_down(Click(float(x_arr[1]), float(y_arr[1])))
[…]
