+++ pyproject.toml
@@ -24,9 +24,9 @@ Cython = "^3.0.10"
 xarray = "^2024.6.0"
 netCDF4 = "^1.7.1"
 opencv-python = "^4.10.0.84"
-img2pdf = "^0.5.1"
 Kivy = "^2.3.0"
 progress = "^1.6"
+tk = "^0.1.0"
 
 [tool.poetry.group.test.dependencies]
 flake8 = "^7.1.0"

[…]
