### 2024-09-19 d947662
Add interactive plotting for all z image. Update docs for transect chain tool
