+++ src/nccut/singletransect.py
@@ -1,174 +1,0 @@
-"""
-Graphics for a single transect whether drawn by user or used to hold data by marker tool.
-
-Draws circle at each user click point. When one point has been clicked, draws line from that
-point to user cursor.
-"""
-
-import kivy.uix as ui
-from kivy.graphics import Color, Ellipse, Line
-from kivy.metrics import dp
-from kivy.uix.label import Label
-from kivy.core.window import Window
-
-
-class SingleTransect(ui.widget.Widget):
-    """
-    Graphics for a single transect whether drawn by user or used to hold data by marker tool.
-
-    Draws circle at each user click point. When one point has been clicked, draws line from that
-    point to user cursor.
-
-    Attributes:
-        line: Line object which holds endpoint data
-        c1: Ellipse object for first point clicked
-        c2: Ellipse object for second point clicked
-        number: Label, reference to number label
-        circles (int): Number of circles drawn, ie points clicked by user
-        home: Reference to root :class:`nccut.homescreen.HomeScreen` instance
-        size: 2 element array of ints, Size of widget
-        pos: 2 element array of ints, Position of widget
-        l_color: kivy.graphics.Color, Color to use for graphics
-        c_size (tuple): 2 element tuple of floats that defines size of circles and other graphics
-        line_width (float): Float, width of lines
-    """
-    def __init__(self, home, **kwargs):
-        """
-        Sets initial settings for transect according to attributes of HomeScreen root
-
-        Args:
-            home: Reference to root :class:`nccut.homescreen.HomeScreen` instance
-        """
-        super(SingleTransect, self).__init__(**kwargs)
-        self.line = None
-        self.c1 = None
-        self.c2 = None
-        self.number = None
-        self.circles = 0
-        self.home = home
-        self.size = self.home.display.size
-        self.pos = self.home.display.pos
-        color = self.home.display.l_col
-        if color == "Blue":
-            self.l_color = Color(0.28, 0.62, 0.86)
-        elif color == "Green":
-            self.l_color = Color(0.39, 0.78, 0.47)
-        elif color == "Orange":
-            self.l_color = Color(0.74, 0.42, 0.13)
-        size = self.home.display.cir_size
-        self.c_size = (dp(size), dp(size))
-        self.line_width = dp(size / 5)
-
-    def update_l_col(self, color):
-        """
-        Updates the line color for the transect and redraws canvas elements
-
-        Args:
-            color (str): New color value: 'Blue', 'Green' or 'Orange'
-        """
-        if color == "Blue":
-            self.l_color = Color(0.28, 0.62, 0.86)
-        elif color == "Green":
-            self.l_color = Color(0.39, 0.78, 0.47)
-        elif color == "Orange":
-            self.l_color = Color(0.74, 0.42, 0.13)
-        graphics = [self.c1, self.c2, self.line]
-        for g in graphics:
-            if g:
-                self.canvas.remove(g)
-        self.canvas.add(self.l_color)
-        for g in graphics:
-            if g:
-                self.canvas.add(g)
-
-    def update_c_size(self, value):
-        """
-        Updates graphic sizes according to the new value.
-
-        Args:
-            value (float): New graphics size
-        """
-        self.line_width = dp(value / 5)
-        self.line.width = self.line_width
-        for c in [self.c1, self.c2]:
-            if c:
-                c.size = (dp(value), dp(value))
-                c.pos = (c.pos[0] + self.c_size[0] / 2 - dp(value) / 2, c.pos[1] + self.c_size[1] / 2 - dp(value) / 2)
-        self.number.font_size = dp(value) * 2
-        self.c_size = (dp(value), dp(value))
-
-    def on_touch_down(self, touch):
-        """
[…]
