### 2024-09-23 2e9c5cb
Fix button image issue
