### 2024-09-25 424f33f
Update file browser to work with macOS
