### 2024-09-25 e334f75
Update file browser to work with macOS
