### 2024-09-26 d939978
Restrict pyobjus dependency to macOS
