+++ pyproject.toml
@@ -27,7 +27,7 @@ opencv-python = "^4.10.0.84"
 Kivy = "^2.3.0"
 progress = "^1.6"
 plyer = "^2.1.0"
-pyobjus = "^1.2.3"
+pyobjus = {version = "^1.2.3", platform = "Darwin"}
 
 [tool.poetry.group.test.dependencies]
 flake8 = "^7.1.0"
