+++ .github/workflows/publish-app.yml
@@ -38,6 +38,7 @@ jobs:
           python -m poetry build
       - name: Commit Version Updates
         run: |
+          git pull
           git config --local user.email "41898282+github-actions[bot]@users.noreply.github.com"
           git config --local user.name "github-actions[bot]"
           git add .
