### 2024-09-30 rchartra/NcCut#15 @platipodium
App unresponsive on Quit()
This is part of https://github.com/openjournals/joss-reviews/issues/7185

The app is quite slow on loading, takes about 2-3 seconds on my M1 laptop.  But worse is that, when I click the Quit button, the windows is not closed but remains as a Zombie process.
