### 2024-09-30 rchartra/NcCut#15 @platipodium
Also hitting the OS system's red button does not close the app.
