### 2024-09-30 rchartra/NcCut#16 @platipodium
User interface not intuitive in interactive session
This issue is part of https://github.com/openjournals/joss-reviews/issues/7185

I followed the examples in the Readme to assess the usability of the interface.  But to me, it is not usable at all. Overall, it does not follow other known interface conventions, not even very old ones like ncview and that is a problem.  As a user testing your examples, I got quickly frustrated by
- work loss
- non-working buttons
- unclear workflow

# Some examples

## top left

As a user, I expect important content on the top left.  But all I get are "view" and "netcdf" buttons that don't do anything if no file is loaded. These buttons are misplaced.
- all buttons that are of no use should be inactive and also visually marked as inactive (greyed out)

## File chooser

It is ok to type or paste a file name into the box.
- Better would be the option to open a file system dialog box letting the user choose (or search for) a file.
- The "Go" button is non-semantic.  It should say "Open" or "Load" in this context.
- white space is not cleaned surrounding pasted file name

## Tools
The Tools square looks too much like a button and I kept hitting it (with frustration as nothing happened)

### Transect/Marker/Quit

These look alike and therefore should do something alike.  But alas: While Transect and Transect Marker buttons open another menu, the Quit button simply quits.   This is confusing
