### 2024-09-30 rchartra/NcCut#16 @platipodium
## Transect

- When I am in Transect mode, both "Transect" and "Transect Marker" buttons appear the same.  How can the user know where she is in the tool?  Very confusing

- transect mode does not do anything whereas Drag mode does.  I don't get it
- Edit mode is fine

## Transect Marker

- It is not intuitive to finish creating a marker transect with the "New Line" button.
- I could not the the width setting to work.
- Also here another "Go" non-semantic button

## Plot

The plot is nice, but: whereas the map is zoomable, the plot is not.  This is again inconsistent UI design
