### 2024-09-30 #7185 @platipodium
@rchartra I had severe trouble with the usability of the interface.  I filed an issue only giving a glimpse into those troubles; you may seek help and other opinions to move this project forward.
