### 2024-09-30 rchartra/NcCut#17 @platipodium
Missing metadata in json output
This relates to https://github.com/openjournals/joss-reviews/issues/7185

When I "save" all data (better naming would be "Export (to json)", the json file does not record the metadata sufficiently.  For GMP (Good Modeling Practice) it is essential to keep, better even augment metadata on every data file creation.

1) It should copy all available metadata from the original data, at least filename (from an image), and at least all global and variable-specific attributes from netCDF

2) It could also add creation timestamp and user name

3) It should add a license and copyright statement, even if empty or with a sensible default.

See also #18
