### 2024-09-30 rchartra/NcCut#15 @rchartra
First, from some of your other issues it appears you are using an outdated version of NcCut. Could you update your version to 0.4.4 and let me know if you're still having problems? I suspect however that some of these issues will persist so I'll investigate this.
