### 2024-09-30 rchartra/NcCut#16 @rchartra
Some of these points have been addressed in the newest version of NcCut 0.4.4:
- Zoomable plot
- Option to use native OS file browser to select a file
- Units are shown for variable when provided by NetCDF file
- New line button for a transect marker has been renamed to New Marker. You can also double click when you draw your last points to finish the marker

Regarding drag mode and transect mode:
- The expected behavior is the following:
  - When you first open a tool, you can drag and zoom but whenever you click, points will be drawn adding to the tool.
  - Selecting drag mode allows you to just drag the image without adding new points.
  - Transect mode closes drag mode, meaning that when you click new points will be drawn and added to the tool
- Can you clarify whether or not you are seeing this behavior, and if so what you find unintuitive about it?

Regarding Marker Width:
- Updating the width will not change previous transects drawn, rather it will set the width for future transects. This is outlined in the documentation and is intentional as it allows for markers with transects of variable width.
- Can you clarify whether or not you are seeing this behavior?

With the remaining issues I see your point, and I'll update the app accordingly.
