### 2024-09-30 rchartra/NcCut#17 @rchartra
I see your point, I'll add a metadata field to the output
