### 2024-10-01 rchartra/NcCut#16 @platipodium
Thanks for tackling this, I see in your Readme that there is a different version.  I will try again and clarify my points above.
