### 2024-10-01 rchartra/NcCut#21 @platipodium
 Error on file chooser: ModuleNotFoundError: No module named 'pyobjus'
Related: https://github.com/openjournals/joss-reviews/issues/7185

I installed via conda/pip/poetry, it seems that the pyobjus dependency is not resolved correctly, though it is present in pyproject.toml

I can install it later with `pip install pyobjus`

* resolve dependencies correctly
* please add an automated test for this issue
