### 2024-10-01 rchartra/NcCut#22 @platipodium
 IndexError: list index out of range. on FileChooser dialog
Related issues
* https://github.com/openjournals/joss-reviews/issues/7185
* #21

This happens when I click the File Chooser (NcCut 0.4.5)

```
 Traceback (most recent call last):
   File "/opt/conda/envs/nccut/bin/nccut", line 6, in <module>
     sys.exit(run())
              ^^^^^
   File "/Users/Lemmen/devel/python/nccut/src/nccut/nccut.py", line 106, in run
     NcCut().run()
   File "/opt/conda/envs/nccut/lib/python3.12/site-packages/kivy/app.py", line 956, in run
     runTouchApp()
   File "/opt/conda/envs/nccut/lib/python3.12/site-packages/kivy/base.py", line 574, in runTouchApp
     EventLoop.mainloop()
   File "/opt/conda/envs/nccut/lib/python3.12/site-packages/kivy/base.py", line 339, in mainloop
     self.idle()
   File "/opt/conda/envs/nccut/lib/python3.12/site-packages/kivy/base.py", line 383, in idle
     self.dispatch_input()
   File "/opt/conda/envs/nccut/lib/python3.12/site-packages/kivy/base.py", line 334, in dispatch_input
     post_dispatch_input(*pop(0))
   File "/opt/conda/envs/nccut/lib/python3.12/site-packages/kivy/base.py", line 263, in post_dispatch_input
     listener.dispatch('on_motion', etype, me)
   File "kivy/_event.pyx", line 731, in kivy._event.EventDispatcher.dispatch
   File "/opt/conda/envs/nccut/lib/python3.12/site-packages/kivy/core/window/__init__.py", line 1709, in on_motion
     self.dispatch('on_touch_down', me)
   File "kivy/_event.pyx", line 731, in kivy._event.EventDispatcher.dispatch
   File "/opt/conda/envs/nccut/lib/python3.12/site-packages/kivy/core/window/__init__.py", line 1726, in on_touch_down
     if w.dispatch('on_touch_down', touch):
        ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
   File "kivy/_event.pyx", line 731, in kivy._event.EventDispatcher.dispatch
   File "/opt/conda/envs/nccut/lib/python3.12/site-packages/kivy/uix/screenmanager.py", line 1210, in on_touch_down
     return super(ScreenManager, self).on_touch_down(touch)
            ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
   File "/opt/conda/envs/nccut/lib/python3.12/site-packages/kivy/uix/widget.py", line 589, in on_touch_down
     if child.dispatch('on_touch_down', touch):
        ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
   File "kivy/_event.pyx", line 731, in kivy._event.EventDispatcher.dispatch
   File "/opt/conda/envs/nccut/lib/python3.12/site-packages/kivy/uix/relativelayout.py", line 306, in on_touch_down
     ret = super(RelativeLayout, self).on_touch_down(touch)
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
   File "/opt/conda/envs/nccut/lib/python3.12/site-packages/kivy/uix/widget.py", line 589, in on_touch_down
     if child.dispatch('on_touch_down', touch):
        ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
   File "kivy/_event.pyx", line 731, in kivy._event.EventDispatcher.dispatch
   File "/opt/conda/envs/nccut/lib/python3.12/site-packages/kivy/uix/widget.py", line 589, in on_touch_down
     if child.dispatch('on_touch_down', touch):
        ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
   File "kivy/_event.pyx", line 731, in kivy._event.EventDispatcher.dispatch
   File "/opt/conda/envs/nccut/lib/python3.12/site-packages/kivy/uix/widget.py", line 589, in on_touch_down
     if child.dispatch('on_touch_down', touch):
        ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
   File "kivy/_event.pyx", line 731, in kivy._event.EventDispatcher.dispatch
   File "/opt/conda/envs/nccut/lib/python3.12/site-packages/kivy/uix/behaviors/button.py", line 151, in on_touch_down
     self.dispatch('on_press')
   File "kivy/_event.pyx", line 727, in kivy._event.EventDispatcher.dispatch
   File "kivy/_event.pyx", line 1307, in kivy._event.EventObservers.dispatch
   File "kivy/_event.pyx", line 1191, in kivy._event.EventObservers._dispatch
   File "/opt/conda/envs/nccut/lib/python3.12/site-packages/kivy/lang/builder.py", line 60, in custom_callback
     exec(__kvlang__.co_value, idmap)
   File "/Users/Lemmen/devel/python/nccut/src/nccut/nccut.kv", line 393, in <module>
     on_press: root.browse()
 ^
   File "/Users/Lemmen/devel/python/nccut/src/nccut/homescreen.py", line 157, in browse
     files = filechooser.open_file(filters=[["Valid Files", "*.png", "*.jpg", "*.jpeg", "*.nc"]])
             ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
   File "/opt/conda/envs/nccut/lib/python3.12/site-packages/plyer/facades/filechooser.py", line 54, in open_file
     return self._file_selection_dialog(mode="open", *args, **kwargs)
            ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
   File "/opt/conda/envs/nccut/lib/python3.12/site-packages/plyer/platforms/macosx/filechooser.py", line 122, in _file_selection_dialog
     return MacFileChooser(**kwargs).run()
            ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
   File "/opt/conda/envs/nccut/lib/python3.12/site-packages/plyer/platforms/macosx/filechooser.py", line 59, in run
     panel = NSOpenPanel.openPanel()
             ^^^^^^^^^^^^^^^^^^^^^^^
   File "pyobjus/pyobjus.pyx", line 504, in pyobjus.ObjcMethod.__call__
   File "pyobjus/pyobjus_conversions.pxi", line 240, in pyobjus.convert_cy_ret_to_py
   File "pyobjus/pyobjus_conversions.pxi", line 181, in pyobjus.convert_to_cy_cls_instance
   File "pyobjus/pyobjus.pyx", line 741, in pyobjus.autoclass
   File "pyobjus/pyobjus.pyx", line 627, in pyobjus.resolve_super_class_methods
   File "pyobjus/pyobjus.pyx", line 543, in pyobjus.class_get_methods
   File "pyobjus/pyobjus.pyx", line 535, in pyobjus.objc_method_to_py
   File "pyobjus/pyobjus.pyx", line 244, in pyobjus.ObjcMethod.__init__
   File "pyobjus/type_enc.pxi", line 13, in pyobjus.parse_signature
   File "pyobjus/type_enc.pxi", line 2, in pyobjus.seperate_encoding
 IndexError: list index out of range
```
