### 2024-10-01 rchartra/NcCut#23 @platipodium
File opener
Related issues

* #21 #22 #16
* https://github.com/openjournals/joss-reviews/issues/7185

As mentioned before #16 the placement of the file chooser on the bottom does not follow logical order (as that would be expected somewhere else).  Also, manually typing in a file path is not conventional, but **I like it**.

There are a number of issues collected here

- [x] problem with MAc file chooser documented as #22
- [x] problem with the dependency #21
- [x] white space pasting problem, pasting 'support/example_4v.nc ' into the box yields an error that can be easily fixed by trimming whitespace around file names.
- [x] when I type in a file name, I want to hit <return> instead of pressing the "Go" button.  But instead, a new line is added leading to "Invalid file name".  Please fix
- [x] I still dislike the "Go" button naming instead of "Load"
