### 2024-10-01 rchartra/NcCut#15 @platipodium
This is fixed in 0.4.5.  App loading with 2 seconds is still a small wait and may annoy users.
