### 2024-10-01 rchartra/NcCut#24 @Huite
Git repository is bloated by committed binary blobs (netcdf, mp4)
Related to: https://github.com/openjournals/joss-reviews/issues/7185

This isn't very relevant for your users, but it is for you and other developers: I just cloned the repo and the first thing I noticed: it takes a while since it's downloading over 400 MB.

If the artifacts are a few MB, I'd say it's perfectly tolerable, but in this case the vast majority is taken up by the netCDFs and mp4 files (once in e.g. `/support/`, but also once again in the git store due to how git works). I know it's somewhat convenient to just check in binary blobs like netCDF files and mp4's, but git isn't made for it. There's stuff like Git LFS but it has downsides/annoyances too. A pragmatic solution without external hosting might be creating a GitHub release specific for your binary content there (obviously it's without version control) and you just use the github URLs in your docs.

I see you're already using `pooch`, but only for your tests? It's convenient for examples as well.
I also see you are using an external link there: `TEST_NC_PATH = pooch.retrieve(url="doi:10.5281/zenodo.12734574/test_nc.nc", known_hash=None)
`
(Adding the hash is worthwhile btw, so you won't get a corrupted file.)
