### 2024-10-01 rchartra/NcCut#25 @Huite
Add a .gitignore and remove dist directory
Related to: https://github.com/openjournals/joss-reviews/issues/7185

You're missing a `.gitignore`, which will insert a lot of meaningless stuff into your git status info. You can find sane defaults at: https://github.com/github/gitignore/blob/main/Python.gitignore

It's a bit big, you can always pare it down, but there's no harm in having a few extra entries.

With regards to the dist dir, it has a temporary nature:
The "dist" directory typically contains built distributions (like .whl files) and source distributions (.tar.gz files) that are generated for uploading to PyPI. These are considered build artifacts rather than source code. Also relates to #24: the `tar.gz` binary blobs inflate your repo size.
