### 2024-10-01 rchartra/NcCut#26 @Huite
Licensing branch is set as default
Related to: https://github.com/openjournals/joss-reviews/issues/7185

Presumably, the main branch should be the default.

![image](https://github.com/user-attachments/assets/f6345095-09e8-4578-87ee-63b70c0d37e1)

Currently, https://github.com/rchartra/NcCut lands on the licensing branch. A clone will then also result in a checkout of that branch rather than main.
