### 2024-10-01 rchartra/NcCut#27 @Huite
Is Cython required?
Related to: https://github.com/openjournals/joss-reviews/issues/7185

The pyproject.toml dependencies mentions Cython: is it really required? I see no Cython code in the project.
And at least on Windows, it seems like all the wheels are available and no compilation is required.

On Windows, Cython is painful because it requires a C compiler and most "non-expert" users will not manage (hence why we distribute built wheels on pypi or use conda).

It might be wise to add a comment in pyproject toml explaining why dependencies are required if they are not imported. E.g xarray will use netCDF4 in open_dataset. I see pyobjus is a bridge with objective-C by kivy, so I can infer kivy needs it on macOS, but an explicit comment is nice.
