### 2024-10-01 rchartra/NcCut#28 @Huite
Is OpenCV required?
Related to: https://github.com/openjournals/joss-reviews/issues/7185

I'm wondering whether you need OpenCV as a relatively heavy dependency. A quick greps tells me:

```python
# in plotpopup
            cv2.imwrite(f_path, img)
...
            img = cv2.flip(cv2.imread(f_path), 0)

# in functions
    c_arr = (np.arange(0, 256) * np.ones((10, 256))).astype(np.uint8).T
    c_bar = cv2.applyColorMap(c_arr, colormap)
    c_bar = cv2.cvtColor(c_bar, cv2.COLOR_BGR2RGB)

# in filedisplay
        out = cv2.addWeighted(data, alpha_c, data, 0, gamma_c)
...
        is_success, img_b = cv2.imencode(".png", img)
...
        c_mapped = cv2.applyColorMap(n_data, self.colormap)
...
        self.cmaps = {"Autumn": cv2.COLORMAP_AUTUMN, "Bone": cv2.COLORMAP_BONE, "Jet": cv2.COLORMAP_JET,
                      "Winter": cv2.COLORMAP_WINTER, "Rainbow": cv2.COLORMAP_RAINBOW, "Ocean": cv2.COLORMAP_OCEAN,
                      "Summer": cv2.COLORMAP_SUMMER, "Spring": cv2.COLORMAP_SPRING, "Cool": cv2.COLORMAP_COOL,
                      "HSV": cv2.COLORMAP_HSV, "Pink": cv2.COLORMAP_PINK, "Hot": cv2.COLORMAP_HOT,
                      "Parula": cv2.COLORMAP_PARULA, "Magma": cv2.COLORMAP_MAGMA, "Inferno": cv2.COLORMAP_INFERNO,
                      "Plasma": cv2.COLORMAP_PLASMA, "Viridis": cv2.COLORMAP_VIRIDIS, "Cividis": cv2.COLORMAP_CIVIDIS,
                      "Twilight": cv2.COLORMAP_TWILIGHT, "Twilight Shifted": cv2.COLORMAP_TWILIGHT_SHIFTED,
                      "Turbo": cv2.COLORMAP_TURBO, "Deep Green": cv2.COLORMAP_DEEPGREEN}
```

My impression is that OpenCV is used very lightly, and the functionality could be provided relatively painlessly by other packages (that are required already):

* Colormaps are available in matplotlib. A more lightweight package like colorcet could also be wortwhile: https://github.com/holoviz/colorcet; Alternatively, you can also just define the matplotlib colormap directly in a python module. Colorcet has the advantage of being a pure Python package, so there's a lot less than can go wrong.
* Matplotlib should be able to save PNGs directly as well; I guess it depends on the backend. Matplotlib can use Pillow https://github.com/python-pillow/Pillow; you can also use Pillow directly with numpy arrays.
* Numpy operations are likely sufficient to replace `applyColorMap` and `addWeighted`, possibly even as one-liners.

Of course, you can choose to maintain OpenCV, but it feels like a non-obvious choice. In that case it's worth documenting and motivating the choice in a local comment.
