### 2024-10-01 rchartra/NcCut#28 @platipodium
Talking about colormaps: there is a set of specific ocean colormaps in the package cmocean.  You might want to add it instead or in addition to the standard colormaps provided above (which do not depend on opencv)
