### 2024-10-01 rchartra/NcCut#29 @Huite
"Save plot to PNG" crashes the application
Related to: https://github.com/openjournals/joss-reviews/issues/7185

When I make a transect chain, click plot, then click "Save plot to PNG", it crashes with the following stacktrace:

```
[ WARN:0@79.017] global loadsave.cpp:241 cv::findDecoder imread_('C:\src\NcCut\support\test'): can't open/read file:
 check file path/integrity
 Traceback (most recent call last):
   File "C:\src\NcCut\.pixi\envs\default\lib\runpy.py", line 197, in _run_module_as_main
     return _run_code(code, main_globals, None,
   File "C:\src\NcCut\.pixi\envs\default\lib\runpy.py", line 87, in _run_code
     exec(code, run_globals)
   File "C:\src\NcCut\.pixi\envs\default\Scripts\nccut.exe\__main__.py", line 7, in <module>
   File "C:\src\NcCut\src\nccut\nccut.py", line 106, in run
     NcCut().run()
   File "C:\src\NcCut\.pixi\envs\default\lib\site-packages\kivy\app.py", line 956, in run
     runTouchApp()
   File "C:\src\NcCut\.pixi\envs\default\lib\site-packages\kivy\base.py", line 574, in runTouchApp
     EventLoop.mainloop()
   File "C:\src\NcCut\.pixi\envs\default\lib\site-packages\kivy\base.py", line 339, in mainloop
     self.idle()
   File "C:\src\NcCut\.pixi\envs\default\lib\site-packages\kivy\base.py", line 383, in idle
     self.dispatch_input()
   File "C:\src\NcCut\.pixi\envs\default\lib\site-packages\kivy\base.py", line 334, in dispatch_input
     post_dispatch_input(*pop(0))
   File "C:\src\NcCut\.pixi\envs\default\lib\site-packages\kivy\base.py", line 263, in post_dispatch_input
     listener.dispatch('on_motion', etype, me)
   File "kivy\\_event.pyx", line 731, in kivy._event.EventDispatcher.dispatch
   File "C:\src\NcCut\.pixi\envs\default\lib\site-packages\kivy\core\window\__init__.py", line 1709, in on_motion
     self.dispatch('on_touch_down', me)
   File "kivy\\_event.pyx", line 731, in kivy._event.EventDispatcher.dispatch
   File "C:\src\NcCut\.pixi\envs\default\lib\site-packages\kivy\core\window\__init__.py", line 1726, in on_touch_dow
n
     if w.dispatch('on_touch_down', touch):
   File "kivy\\_event.pyx", line 731, in kivy._event.EventDispatcher.dispatch
   File "C:\src\NcCut\.pixi\envs\default\lib\site-packages\kivy\uix\popup.py", line 236, in on_touch_down
     return super(Popup, self).on_touch_down(touch)
   File "C:\src\NcCut\.pixi\envs\default\lib\site-packages\kivy\uix\modalview.py", line 266, in on_touch_down
     super().on_touch_down(touch)
   File "C:\src\NcCut\.pixi\envs\default\lib\site-packages\kivy\uix\widget.py", line 589, in on_touch_down
     if child.dispatch('on_touch_down', touch):
   File "kivy\\_event.pyx", line 731, in kivy._event.EventDispatcher.dispatch
   File "C:\src\NcCut\.pixi\envs\default\lib\site-packages\kivy\uix\widget.py", line 589, in on_touch_down
     if child.dispatch('on_touch_down', touch):
   File "kivy\\_event.pyx", line 731, in kivy._event.EventDispatcher.dispatch
   File "C:\src\NcCut\.pixi\envs\default\lib\site-packages\kivy\uix\widget.py", line 589, in on_touch_down
     if child.dispatch('on_touch_down', touch):
   File "kivy\\_event.pyx", line 731, in kivy._event.EventDispatcher.dispatch
   File "C:\src\NcCut\.pixi\envs\default\lib\site-packages\kivy\uix\widget.py", line 589, in on_touch_down
     if child.dispatch('on_touch_down', touch):
   File "kivy\\_event.pyx", line 731, in kivy._event.EventDispatcher.dispatch
   File "C:\src\NcCut\.pixi\envs\default\lib\site-packages\kivy\uix\widget.py", line 589, in on_touch_down
     if child.dispatch('on_touch_down', touch):
   File "kivy\\_event.pyx", line 731, in kivy._event.EventDispatcher.dispatch
   File "C:\src\NcCut\.pixi\envs\default\lib\site-packages\kivy\uix\behaviors\button.py", line 151, in on_touch_down
     self.dispatch('on_press')
   File "kivy\\_event.pyx", line 727, in kivy._event.EventDispatcher.dispatch
   File "kivy\\_event.pyx", line 1307, in kivy._event.EventObservers.dispatch
   File "kivy\\_event.pyx", line 1231, in kivy._event.EventObservers._dispatch
   File "C:\src\NcCut\src\nccut\plotpopup.py", line 189, in <lambda>
     png_btn.bind(on_press=lambda x: self.file_input('png'))
   File "C:\src\NcCut\src\nccut\plotpopup.py", line 313, in file_input
     self.download_png_plot(fpath)
   File "C:\src\NcCut\src\nccut\plotpopup.py", line 329, in download_png_plot
     cv2.imwrite(f_path, img)
 cv2.error: OpenCV(4.10.0) D:\a\opencv-python\opencv-python\opencv\modules\imgcodecs\src\loadsave.cpp:798: error: (-
215:Assertion failed) !_img.empty() in function 'cv::imwrite'
```

I haven't carefully looked why it's failing here (could be the case that my installation isn't quite right), but I can note that you'd want to employ a bit more defensive programming here. Whenever an operation fails  -- for whatever reason, let's say I'm out of disk space -- it shouldn't crash the app.
In this case it's ease to solve with a `try except`. Ideally, you then provide user feedback as well as you do for invalid netCDF paths and the like.

Qt even has a specific widget for it: https://doc.qt.io/qt-6/qerrormessage.html#:~:text=For%20production%20applications,%20the%20class

I don't see exactly the same thing for kivy, but a basic popup will probably suffice.
