### 2024-10-01 rchartra/NcCut#29 @Huite
I might've found the issue. The file picker states "save as type: *.png", but apparently `.png` isn't suffixed properly. OpenCV then probably doesn't recognize what to do.

This seems like an issue in the plyer filechooser `save_file` method (maybe platform dependent). I recommend using `pathlib.Path` call `with_suffix` as needed.

In terms of OpenCV, this'll do the trick too:

```python
            from PIL import Image as PILImage

            texture = self.plot.texture
            pil_image = PILImage.frombytes('RGBA', texture.size, texture.pixels)
            pil_image.save(f_path, 'png')
```

Happens to be more robust in this specific case, because this `save` method explicitly requires the extension, so it creates a valid PNG even if the name doesn't have the right suffix.
