### 2024-10-02 rchartra/NcCut#26 @rchartra
I think this happened when I renamed the "master" branch to "main". I've updated it so now "main" is now the default branch.
