### 2024-10-03 dbe8649
Fix save plot on PNG bug. Clearer workflow. Error popups.
