+++ src/nccut/markerwidth.py
@@ -23,7 +23,7 @@ class MarkerWidth(ui.boxlayout.BoxLayout):
 
     Attributes:
         txt: kivy.uix.textinput.TextInput widget where user enters their desired width
-        btn: 'Go' Button
+        btn: 'Set' Button
         marker: Current :class:`nccut.multimarker.MultiMarker` instance
         min (int): Minimum width allowed
         max (int): Maximum width allowed
@@ -37,7 +37,8 @@ class MarkerWidth(ui.boxlayout.BoxLayout):
         """
         super(MarkerWidth, self).__init__(**kwargs)
         self.txt = TextInput(hint_text="Width", size_hint=(.7, 1))
-        self.btn = func.RoundedButton(text="Go", size_hint=(.3, 1), font_size=marker.home.font)
+        self.font_size = marker.home.font
+        self.btn = func.RoundedButton(text="Set", size_hint=(.3, 1), font_size=self.font_size)
         self.btn.bind(on_press=lambda x: self.update())
         self.add_widget(self.txt)
         self.add_widget(self.btn)
