+++ src/nccut/marker.py
@@ -205,10 +205,10 @@ class Marker(ui.widget.Widget):
             self.transects = self.transects[:-1]
         else:
             # Remove plot and width buttons from sidebar if last point of the marker
-            if self.parent.dbtn in self.home.display.current:
-                self.home.display.current.remove(self.parent.dbtn)
-            if self.parent.width_w in self.home.display.current:
-                self.home.display.current.remove(self.parent.width_w)
+            if self.parent.dbtn in self.home.display.tool_action_widgets:
+                self.home.display.remove_from_tool_action_widgets(self.parent.dbtn)
+            if self.parent.width_w in self.home.display.tool_action_widgets:
+                self.home.display.remove_from_tool_action_widgets(self.parent.width_w)
             self.remove_widget(self.children[0])
             # Stop drawing line between last point and cursor
             Window.unbind(mouse_pos=self.draw_line)
