### 2024-10-03 e1106dd
Fix save plot on PNG bug. Clearer workflow. Error popups.
