+++ tests/test_gui.py
@@ -13,6 +13,7 @@ import copy
 import pooch
 import json
 import cv2
+import tempfile
 import numpy as np
 from functools import partial
 from kivy.metrics import dp
@@ -20,7 +21,6 @@ from kivy.clock import Clock
 from kivy.uix.button import Button
 from kivy.uix.checkbox import CheckBox
 from nccut.multimarker import Click, marker_find
-from nccut.markerwidth import MarkerWidth
 import nccut.functions as functions
 from nccut.nccut import NcCut
 
@@ -62,7 +62,7 @@ def get_app():
 
 def load_2d_nc(variable):
     run_app.home.ids.file_in.text = SUPPORT_FILE_PATH + "example_4v.nc"
-    run_app.home.go_btn()
+    run_app.home.open_btn()
     popup = run_app.home.nc_popup
     popup.var_select.text = variable
     popup.x_select.text = "x"
@@ -73,7 +73,7 @@ def load_2d_nc(variable):
 
 def load_3d_nc(z_val):
     run_app.home.ids.file_in.text = SUPPORT_FILE_PATH + "example_3d.nc"
-    run_app.home.go_btn()
+    run_app.home.open_btn()
     popup = run_app.home.nc_popup
     popup.var_select.text = "Theta"
     popup.x_select.text = "i"
@@ -85,7 +85,7 @@ def load_3d_nc(z_val):
 
 
 def select_sidebar_button(text):
-    sidebar = run_app.home.ids.sidebar
+    sidebar = run_app.home.ids.dynamic_sidebar
     but = next((but for but in sidebar.children if isinstance(but, Button) and but.text == text), None)
     but.dispatch('on_press')
     but.dispatch('on_release')
@@ -101,18 +101,19 @@ class Test(unittest.TestCase):
         """
         Test input file name rules
         """
-        run_app.home.ids.file_in.text = "*$%&@! "
-        run_app.home.go_btn()
+        run_app.home.ids.file_in.text = " *$%&@! "
+        run_app.home.open_btn()
         self.assertEqual(run_app.home.children[0].text, "Invalid File Name",
                          "File names with irregular characters are invalid")
+        self.assertEqual(run_app.home.ids.file_in.text, "*$%&@!", "Whitespace not removed from file entry")
         run_app.home.ids.file_in.text = ""
-        run_app.home.go_btn()
+        run_app.home.open_btn()
         self.assertEqual(run_app.home.children[0].text, "Invalid File Name", "Empty file names are invalid")
         run_app.home.ids.file_in.text = SUPPORT_FILE_PATH + "project_example.json"
-        run_app.home.go_btn()
+        run_app.home.open_btn()
         self.assertEqual(run_app.home.children[0].text, "Unsupported File Type", "No unaccepted file types")
         run_app.home.ids.file_in.text = "teacup.jpg"
-        run_app.home.go_btn()
+        run_app.home.open_btn()
         self.assertEqual(run_app.home.children[0].text, "File Not Found", "File must exist")
 
     def test_tool_cleanup(self):
@@ -120,35 +121,38 @@ class Test(unittest.TestCase):
         Check that viewer window is clean after removing tools
         """
         run_app.home.ids.file_in.text = SUPPORT_FILE_PATH + "example.jpg"
-        run_app.home.go_btn()
+        run_app.home.open_btn()
         self.assertEqual(run_app.home.file_on, True, "File was not loaded")
 
         # Run for all tools in sidebar
-        og_side = copy.copy(run_app.home.ids.sidebar.children)
+        og_side = ["Transect Marker", "Transect Chain"]
         og_img = copy.copy(run_app.home.display.children[0].children)
-        for item in run_app.home.ids.sidebar.children:
[…]
