+++ src/nccut/plotpopup.py
@@ -14,19 +14,18 @@ from kivy.uix.popup import Popup
 from kivy.uix.label import Label
 from kivy.uix.checkbox import CheckBox
 from kivy.uix.dropdown import DropDown
+from kivy.uix.widget import Widget
 from plyer import filechooser
 import nccut.functions as func
 from nccut.plotwindow import PlotWindow
 from kivy.core.image import Image as CoreImage
 import matplotlib.pyplot as plt
 from PIL import Image as im
-from PIL import ImageOps as imo
 import numpy as np
 import copy
-import os
 import pandas as pd
 import json
-import cv2
+import tempfile
 
 
 class BackgroundDropDown(DropDown):
@@ -174,27 +173,31 @@ class PlotPopup(Popup):
 
         # Saving Data/Plotting buttons
 
-        self.buttons = ui.boxlayout.BoxLayout(orientation='horizontal', size_hint=(1, .1), spacing=dp(10))
+        self.buttons_bar = ui.boxlayout.BoxLayout(orientation='horizontal', size_hint=(1, .1))
+        self.buttons = ui.boxlayout.BoxLayout(orientation='horizontal', size_hint=(0.8, 1), spacing=dp(10))
 
         self.f_m = 0.8
-        a_data_btn = func.RoundedButton(text="Save All Data", size_hint=(.15, 1),
+        a_data_btn = func.RoundedButton(text="Export All Data", size_hint=(.15, 1),
                                              font_size=self.home.font * self.f_m)
         a_data_btn.bind(on_press=lambda x: self.file_input('a_data'))
 
-        s_data_btn = func.RoundedButton(text="Save Selected Data", size_hint=(.15, 1),
+        s_data_btn = func.RoundedButton(text="Export Selected Data", size_hint=(.15, 1),
                                         font_size=self.home.font * self.f_m)
         s_data_btn.bind(on_press=lambda x: self.file_input('s_data'))
 
-        png_btn = func.RoundedButton(text='Save Plot to PNG', size_hint=(.15, 1), font_size=self.home.font * self.f_m)
+        png_btn = func.RoundedButton(text='Export Plot to PNG', size_hint=(.15, 1), font_size=self.home.font * self.f_m)
         png_btn.bind(on_press=lambda x: self.file_input('png'))
 
-        pdf_btn = func.RoundedButton(text='Save Plot to PDF', size_hint=(.15, 1), font_size=self.home.font * self.f_m)
+        pdf_btn = func.RoundedButton(text='Export Plot to PDF', size_hint=(.15, 1), font_size=self.home.font * self.f_m)
         pdf_btn.bind(on_press=lambda x: self.file_input('pdf'))
 
         self.buttons.add_widget(a_data_btn)
         self.buttons.add_widget(s_data_btn)
         self.buttons.add_widget(png_btn)
         self.buttons.add_widget(pdf_btn)
+
+        self.buttons_bar.add_widget(self.buttons)
+        self.buttons_bar.add_widget(Widget(size_hint=(0.1, 1)))
 
         # Transect selection
         t_box = ui.boxlayout.BoxLayout(size_hint=(1, 0.2), spacing=dp(5))
@@ -248,7 +251,7 @@ class PlotPopup(Popup):
                 zp_box.add_widget(zp_btn)
                 sidebar.add_widget(zp_box)
 
-                allz_btn = func.RoundedButton(text="Save All Z Data", size_hint=(.2, 1),
+                allz_btn = func.RoundedButton(text="Export All Z Data", size_hint=(.2, 1),
                                               font_size=self.home.font * self.f_m)
                 allz_btn.bind(on_press=lambda x: self.file_input('all_z'))
[…]

[…]
