### 2024-10-03 rchartra/NcCut#25 @rchartra
I see. I've removed the dist directory and included a .gitignore
