### 2024-10-03 rchartra/NcCut#29 @rchartra
Thank you for the tips and finding this bug! I've resolved the bug and implemented tempfiles wherever I was creating files temporarily. I've also switched over to PIL as part of #28 . The changes are in the ui_clarity branch for now. I'll include these fixes in the next release as well as fixes to the other issues.
