+++ tests/test_gui.py
@@ -25,8 +25,11 @@ import nccut.functions as functions
 from nccut.nccut import NcCut
 
 os.environ['KIVY_GL_BACKEND'] = 'angle_sdl2'
-SUPPORT_FILE_PATH = "support/"
 TEST_NC_PATH = pooch.retrieve(url="doi:10.5281/zenodo.12734574/test_nc.nc", known_hash=None)
+EXAMPLE_JPG_PATH = pooch.retrieve(url="doi:10.5281/zenodo.13883476/example.jpg", known_hash=None)
+EXAMPLE_3D_PATH = pooch.retrieve(url="doi:10.5281/zenodo.13883476/example_3d.nc", known_hash=None)
+EXAMPLE_4V_PATH = pooch.retrieve(url="doi:10.5281/zenodo.13883476/example_4v.nc", known_hash=None)
+PROJECT_EXAMPLE_PATH = pooch.retrieve(url="doi:10.5281/zenodo.13883476/project_example.json", known_hash=None)
 
 
 class AppInfo:
@@ -61,7 +64,7 @@ def get_app():
 
 
 def load_2d_nc(variable):
-    run_app.home.ids.file_in.text = SUPPORT_FILE_PATH + "example_4v.nc"
+    run_app.home.ids.file_in.text = EXAMPLE_4V_PATH
     run_app.home.go_btn()
     popup = run_app.home.nc_popup
     popup.var_select.text = variable
@@ -72,7 +75,7 @@ def load_2d_nc(variable):
 
 
 def load_3d_nc(z_val):
-    run_app.home.ids.file_in.text = SUPPORT_FILE_PATH + "example_3d.nc"
+    run_app.home.ids.file_in.text = EXAMPLE_3D_PATH
     run_app.home.go_btn()
     popup = run_app.home.nc_popup
     popup.var_select.text = "Theta"
@@ -108,7 +111,7 @@ class Test(unittest.TestCase):
         run_app.home.ids.file_in.text = ""
         run_app.home.go_btn()
         self.assertEqual(run_app.home.children[0].text, "Invalid File Name", "Empty file names are invalid")
-        run_app.home.ids.file_in.text = SUPPORT_FILE_PATH + "project_example.json"
+        run_app.home.ids.file_in.text = PROJECT_EXAMPLE_PATH
         run_app.home.go_btn()
         self.assertEqual(run_app.home.children[0].text, "Unsupported File Type", "No unaccepted file types")
         run_app.home.ids.file_in.text = "teacup.jpg"
@@ -119,7 +122,7 @@ class Test(unittest.TestCase):
         """
         Check that viewer window is clean after removing tools
         """
-        run_app.home.ids.file_in.text = SUPPORT_FILE_PATH + "example.jpg"
+        run_app.home.ids.file_in.text = EXAMPLE_JPG_PATH
         run_app.home.go_btn()
         self.assertEqual(run_app.home.file_on, True, "File was not loaded")
 
@@ -146,7 +149,7 @@ class Test(unittest.TestCase):
         og_side = copy.copy(run_app.home.ids.sidebar.children)
 
         # Open a file and use a tool
-        run_app.home.ids.file_in.text = SUPPORT_FILE_PATH + "example_4v.nc"
+        run_app.home.ids.file_in.text = EXAMPLE_4V_PATH
         run_app.home.go_btn()
         load_2d_nc("Vorticity")
         self.assertEqual(len(run_app.home.color_bar_box.children), 1, "Colorbar was not Added")
@@ -163,7 +166,7 @@ class Test(unittest.TestCase):
         for i in range(len(incs)):
             tool.on_touch_down(Click(float(x_arr[i]), float(y_arr[i])))
 
-        run_app.home.ids.file_in.text = SUPPORT_FILE_PATH + "example.jpg"
[…]
