+++ REUSE.toml
@@ -17,9 +17,3 @@ SPDX-License-Identifier = "CC0-1.0"
 path = "support/*"
 SPDX-FileCopyrightText = "2024 University of Washington"
 SPDX-License-Identifier = "CC0-1.0"
-
-[[annotations]]
-
-path = "support/example.jpg"
-SPDX-FileCopyrightText = "Credit: ESA/Sentinel-1, Unaltered"
-SPDX-License-Identifier = "CC-BY-SA-3.0-IGO"
