+++ REUSE.toml
@@ -11,9 +11,3 @@ SPDX-License-Identifier = "BSD-3-Clause"
 path = "images/*"
 SPDX-FileCopyrightText = "2024 University of Washington"
 SPDX-License-Identifier = "CC0-1.0"
-
-[[annotations]]
-
-path = "support/*"
-SPDX-FileCopyrightText = "2024 University of Washington"
-SPDX-License-Identifier = "CC0-1.0"
