### 2024-10-03 rchartra/NcCut#24 @rchartra
Oh yikes.... I'll definitely work on that. I've moved the example files over to zenodo and removed the example video for now. I'll need to clean the files from my repository history though, so this issue isn't resolved yet.
