### 2024-10-03 rchartra/NcCut#28 @rchartra
Yea the only main purpose it serves is the colormaps. I'll replace that with a lighter library.
