### 2024-10-04 rchartra/NcCut#23 @rchartra
Thanks for finding these bugs, I'll get them sorted out.
