### 2024-10-04 733b8cf
Settings bar only visible when accessible
