### 2024-10-04 cdeac45
Settings bar only visible when accessible
