+++ src/nccut/homescreen.py
@@ -23,6 +23,7 @@ from pathlib import Path
 from plyer import filechooser
 from nccut.filedisplay import FileDisplay
 from nccut.netcdfconfig import NetCDFConfig
+from nccut.settingsbar import SettingsBar
 
 
 class HomeScreen(Screen):
@@ -39,6 +40,12 @@ class HomeScreen(Screen):
         display: FileDisplay object (draggable image) created when a file is loaded
         nc_popup: Reference to NetCDF configuration popup
         file: File if one was given on start up from command line, otherwise None
+        color_bar_box: BoxLayout containing colorbar and related graphics
+        cb_bg: Background for colorbar box
+        netcdf_info: Label object containing information about NetCDF file if it has necessary attributes
+        sidebar_label: Label object for title of dynamic sidebar
+        sidebar_spacer: Spacer to fill any remaining area in dynamic sidebar not filled by widgets
+        settings_bar: SettingsBar object holding view manipulation buttons and NetCDF menu
     """
     def __init__(self, btn_img_path, file=None, **kwargs):
         """
@@ -69,6 +76,8 @@ class HomeScreen(Screen):
         # Dynamic sidebar
         self.sidebar_label = self.ids.sidebar_label
         self.sidebar_spacer = self.ids.dynamic_spacer
+        # Settings bar
+        self.settings_bar = SettingsBar(self.font)
 
     def populate_dynamic_sidebar(self, elements, sidebar_label):
         """
@@ -108,7 +117,7 @@ class HomeScreen(Screen):
         if not self.loaded and self.ids.view_box.size[0] / kivy.core.window.Window.size[0] >= 0.75:
             if self.file:
                 self.ids.file_in.text = str(self.file)
-                self.open_btn()
+                self.load_btn()
                 if str(self.file)[-3:] == ".nc":
                     self.nc_popup.load.dispatch("on_press")
             self.loaded = True
@@ -126,12 +135,12 @@ class HomeScreen(Screen):
         for i in range(len(dsl.children)):
             dsl.children[i].font_size = font
         self.font = font
-        self.ids.netcdf_btn.font_size = font
         self.netcdf_info.font_size = font
+        self.settings_bar.font_adapt(font)
         if self.file_on:
             self.display.font_adapt(font)
 
-    def open_btn(self):
+    def load_btn(self):
         """
         If file name is valid and exists, load image or NetCDF File.
 
@@ -156,6 +165,8 @@ class HomeScreen(Screen):
                     # Creates interactive image from .jpg/.png/.jpeg files
                     self.display = FileDisplay(home=self, f_config={"image": str(file)})
                     self.ids.view.add_widget(self.display)
+                    if self.settings_bar.parent is None:
+                        self.ids.settings_bar.add_widget(self.settings_bar)
                     self.file_on = True
                 else:
                     func.alert("Unsupported File Type", self)
@@ -175,7 +186,7 @@ class HomeScreen(Screen):
         else:
             text = files[0]
         self.ids.file_in.text = text
-        self.open_btn()
+        self.load_btn()
 
     def load_colorbar_and_info(self, colorbar, config):
         """
@@ -225,6 +236,10 @@ class HomeScreen(Screen):
         """
         self.display = FileDisplay(home=self, f_config={"netcdf": config})
         self.ids.view.add_widget(self.display)
+        if self.settings_bar.parent is None:
+            self.ids.settings_bar.add_widget(self.settings_bar)
+        self.settings_bar.add_netcdf_button()
+        self.settings_bar.font_adapt(self.font)
         self.file_on = True
 
     def clean_file(self):
@@ -239,8 +254,11 @@ class HomeScreen(Screen):
[…]
