+++ pyproject.toml
@@ -22,12 +22,12 @@ scipy = "^1.13.1"
 matplotlib = "^3.9.0"
 Cython = "^3.0.10"
 xarray = "^2024.6.0"
-netCDF4 = "^1.7.1"
+netCDF4 = "^1.7.1" # Necessary for xarray
 opencv-python = "^4.10.0.84"
 Kivy = "^2.3.0"
 progress = "^1.6"
 plyer = "^2.1.0"
-pyobjus = {version = "^1.2.3", platform = "Darwin"}
+pyobjus = {version = "^1.2.3", platform = "darwin"} # Necessary for native file browser on MacOS
 
 [tool.poetry.group.test.dependencies]
 flake8 = "^7.1.0"
