+++ .github/workflows/python-app.yml
@@ -68,6 +68,26 @@ jobs:
     - name: Test with unittest
       run: |
         python -m poetry run python -m pytest
+
+  mac_deps_test:
+    runs_on: macos-latest
+    steps:
+      - uses: actions/checkout@v4
+      - name: Set up Python 3.x
+        uses: actions/setup-python@v5
+        with:
+          python-version: ${{ matrix.version }}
+      - name: Install Poetry
+        run: |
+          python -m pip install --upgrade pip
+          python -m pip install poetry
+      - name: Install dependencies
+        run: |
+          python -m poetry install --with test
+      - name: Test pyobjus install
+        run: |
+          python -c "import pyobjus"
+
   licenses:
     runs-on: ubuntu-latest
     steps:
