### 2024-10-05 067d015
Better plot popup menu
