+++ src/nccut/plotpopup.py
@@ -7,6 +7,7 @@ Manages the saving of the plots and user selected data.
 
 import io
 import kivy.uix as ui
+from kivy.lang import Builder
 from kivy.graphics import Color, Rectangle
 from kivy.metrics import dp
 from kivy.uix.button import Button
@@ -14,7 +15,6 @@ from kivy.uix.popup import Popup
 from kivy.uix.label import Label
 from kivy.uix.checkbox import CheckBox
 from kivy.uix.dropdown import DropDown
-from kivy.uix.widget import Widget
 from plyer import filechooser
 import nccut.functions as func
 from nccut.plotwindow import PlotWindow
@@ -25,6 +25,11 @@ import numpy as np
 import copy
 import pandas as pd
 import json
+import pathlib
+
+
+KV_FILE_PATH = pathlib.Path(__file__).parent.resolve() / "plotpopup.kv"
+Builder.load_file(str(KV_FILE_PATH))
 
 
 class BackgroundDropDown(DropDown):
@@ -162,58 +167,16 @@ class PlotPopup(Popup):
 
         self.plot = ui.image.Image(source="", texture=CoreImage(io.BytesIO(temp.read()), ext="png").texture,
                                    size_hint=(0.6, 1), fit_mode="contain")
-        # Plot
-        self.plotting = ui.boxlayout.BoxLayout(spacing=dp(20), size_hint=(1, 0.9))
-        self.plotting.add_widget(self.plot)
-        self.title = "Plotting Menu"
-        self.content = ui.boxlayout.BoxLayout(orientation='vertical', spacing=dp(15), padding=dp(10))
-        self.size_hint = (0.8, 0.8)
-        sidebar = ui.boxlayout.BoxLayout(orientation='vertical', size_hint=(0.4, 1), padding=dp(10), spacing=dp(20))
-
-        # Saving Data/Plotting buttons
-
-        self.buttons_bar = ui.boxlayout.BoxLayout(orientation='horizontal', size_hint=(1, .1))
-        self.buttons = ui.boxlayout.BoxLayout(orientation='horizontal', size_hint=(0.8, 1), spacing=dp(10))
-
+        self.ids.plotting.add_widget(self.plot, len(self.ids.plotting.children))
         self.f_m = 0.8
-        a_data_btn = func.RoundedButton(text="Export All Data", size_hint=(.15, 1),
-                                             font_size=self.home.font * self.f_m)
-        a_data_btn.bind(on_press=lambda x: self.file_input('a_data'))
-
-        s_data_btn = func.RoundedButton(text="Export Selected Data", size_hint=(.15, 1),
-                                        font_size=self.home.font * self.f_m)
-        s_data_btn.bind(on_press=lambda x: self.file_input('s_data'))
-
-        png_btn = func.RoundedButton(text='Export Plot to PNG', size_hint=(.15, 1), font_size=self.home.font * self.f_m)
-        png_btn.bind(on_press=lambda x: self.file_input('png'))
-
-        pdf_btn = func.RoundedButton(text='Export Plot to PDF', size_hint=(.15, 1), font_size=self.home.font * self.f_m)
-        pdf_btn.bind(on_press=lambda x: self.file_input('pdf'))
-
-        self.buttons.add_widget(a_data_btn)
-        self.buttons.add_widget(s_data_btn)
-        self.buttons.add_widget(png_btn)
-        self.buttons.add_widget(pdf_btn)
-
-        self.buttons_bar.add_widget(self.buttons)
-        self.buttons_bar.add_widget(Widget(size_hint=(0.1, 1)))
+        self.widgets_with_text = [self.ids.sel_transects_label, self.ids.sel_transects_btn]
 
         # Transect selection
[…]
