### 2024-10-05 6e6882f
Better plot popup menu
