+++ src/nccut/markerwidth.py
@@ -11,6 +11,7 @@ transect marker tool.
 
 import kivy.uix as ui
 from kivy.uix.textinput import TextInput
+from kivy.metrics import dp
 import nccut.functions as func
 
 
@@ -22,6 +23,7 @@ class MarkerWidth(ui.boxlayout.BoxLayout):
     transect marker tool.
 
     Attributes:
+        spacing: The spacing between widgets in the layout
         txt: kivy.uix.textinput.TextInput widget where user enters their desired width
         btn: 'Set' Button
         marker: Current :class:`nccut.multimarker.MultiMarker` instance
@@ -36,6 +38,7 @@ class MarkerWidth(ui.boxlayout.BoxLayout):
             marker: Active :class:`nccut.multimarker.MultiMarker` instance
         """
         super(MarkerWidth, self).__init__(**kwargs)
+        self.spacing = dp(8)
         self.txt = TextInput(hint_text="Width", size_hint=(.7, 1))
         self.font_size = marker.home.font
         self.btn = func.RoundedButton(text="Set", size_hint=(.3, 1), font_size=self.font_size)
