### 2024-10-09 9aebba6
Fix weak ref error
