### 2024-10-09 03f01f0
Remove cv2 as a dependency
