### 2024-10-09 17d0ebb
Remove cv2 as a dependency
