+++ src/nccut/filedisplay.py
@@ -17,9 +17,11 @@ import kivy.uix as ui
 from kivy.core.image import Image as CoreImage
 from kivy.metrics import dp
 from PIL import Image as im
+from PIL import ImageEnhance
 import numpy as np
-import cv2
+import matplotlib.pyplot as plt
 import io
+import warnings
 import nccut.functions as func
 from nccut.multimarker import MultiMarker
 from nccut.multichain import MultiChain
@@ -105,15 +107,8 @@ class FileDisplay(ScatterLayout):
         self.l_col = "Blue"
         self.cir_size = 5
 
-        self.cmaps = {"Autumn": cv2.COLORMAP_AUTUMN, "Bone": cv2.COLORMAP_BONE, "Jet": cv2.COLORMAP_JET,
-                      "Winter": cv2.COLORMAP_WINTER, "Rainbow": cv2.COLORMAP_RAINBOW, "Ocean": cv2.COLORMAP_OCEAN,
-                      "Summer": cv2.COLORMAP_SUMMER, "Spring": cv2.COLORMAP_SPRING, "Cool": cv2.COLORMAP_COOL,
-                      "HSV": cv2.COLORMAP_HSV, "Pink": cv2.COLORMAP_PINK, "Hot": cv2.COLORMAP_HOT,
-                      "Parula": cv2.COLORMAP_PARULA, "Magma": cv2.COLORMAP_MAGMA, "Inferno": cv2.COLORMAP_INFERNO,
-                      "Plasma": cv2.COLORMAP_PLASMA, "Viridis": cv2.COLORMAP_VIRIDIS, "Cividis": cv2.COLORMAP_CIVIDIS,
-                      "Twilight": cv2.COLORMAP_TWILIGHT, "Twilight Shifted": cv2.COLORMAP_TWILIGHT_SHIFTED,
-                      "Turbo": cv2.COLORMAP_TURBO, "Deep Green": cv2.COLORMAP_DEEPGREEN}
-        self.colormap = self.cmaps['Viridis']
+        self.cmaps = plt.colormaps()[:87]
+        self.colormap = 'viridis'
 
         # Initial Sidebar Widgets
         self.transect_chain_btn = func.RoundedButton(text="Transect Chain", size_hint=(1, 0.1),
@@ -180,6 +175,7 @@ class FileDisplay(ScatterLayout):
         """
         if self.f_type == "netcdf":
             self.netcdf_to_image()
+            self.byte.seek(0)
             self.im = CoreImage(self.byte, ext='png')
             self.size = self.im.size
         elif self.f_type == "image":
@@ -319,11 +315,15 @@ class FileDisplay(ScatterLayout):
                 self.tool.update_c_size(float(value))
         elif setting == "contrast":
             if self.f_type == "netcdf":
-                self.contrast = float(value)
+                v = float(value) / 20
+                if v < 0:
+                    self.contrast = 1 + v
+                else:
+                    self.contrast = 1 + v * 2
                 self.update_netcdf()
         elif setting == "colormap":
             if self.f_type == "netcdf":
-                self.colormap = self.cmaps[value]
+                self.colormap = value
                 self.update_netcdf()
         elif setting == "variable":
             if self.f_type == "netcdf":
@@ -339,6 +339,7 @@ class FileDisplay(ScatterLayout):
         Reload netcdf image when netcdf data is changed.
         """
         self.netcdf_to_image()
+        self.byte.seek(0)
         self.im = CoreImage(self.byte, ext='png')
         self.size = self.im.size
         self.img.texture = self.im.texture
@@ -348,42 +349,26 @@ class FileDisplay(ScatterLayout):
         """
         Creates image from NetCDF dataset defined in :attr:`nccut.filedisplay.FileDisplay.f_config`
 
[…]
