+++ README.md
@@ -73,40 +73,43 @@ NcCut().run()
 
 ## Loading a File
 
-1.  To load an image file or NetCDF file type the relative or absolute file path to the file into the file entry text box and select "Go"
+1.  To load an image file or NetCDF file either select "Browse" to locate the file or you can type the relative or absolute file path to the file into the file entry text box and select "Load".
     * If you are loading a NetCDF file a popup window will appear with additional selections:
         * Select which variable from your file you would like to see.
         * Confirm or adjust which dimensions you would like to use as the X and Y axes for your variable.
         * If your variable has data in a third dimension select the Z dimension and an initial Z value to display. You will be able to toggle between z values later.
-        * Select "Go" to load the selected data
+        * Select "Load" to load the selected data
 
 
 2. You can scroll to zoom in and out of the image and can click and drag the image to move it around.
-3. From the "View" menu in the settings bar you can rotate or flip the image as well as change graphic settings for the tools.
+3. From the settings bar you can rotate or flip the image as well as change graphic settings for the tools.
 4. If you are loading a NetCDF file, from the "NetCDF" menu in the settings bar you can change which variable or z value you'd like to see as well as change the color map or contrast of the image.
+   * Changing the contrast and colormap only affects the displayed image. Transect data is taken from the NetCDF file itself.
 
 ## Tools
-* To clear widgets from the viewer at any point click any of the tool buttons
-### Transects
-* This tool will make multiple transects between two points
-
-1. Hit "Transect" to enter transect mode
-2. Click two points you'd like to make a transect between
-3. Repeat for as many transects as you'd like
-4. When done, select the "Plot" button and a popup will appear with a plot of all transects and downloading options.
+* NcCut has two types of tools for making transects.
+### Transect Chain
+* Using this tool chains of transects can be drawn. Transects will be taken along the line segments drawn between the clicked points. Multiple such "Chains" can be drawn out at once before being plotted.
 
 
-### Transect Markers
-* This tool allows you to click along a feature and have transects automatically made perpendicular to the feature. 
-* You can mark out multiple features and save them all together as a single "project". 
-* You can upload previous projects to edit or continue them.
+1. Hit "Transect Chain" to enter transect mode
+2. Click points along the feature you'd like to make transects along. Transects will be made between the points you click.
+    * Select "Drag Mode" from the actions sidebar to drag the image without selecting points, and select "Transect Mode" to go back.
+    * Select "Edit Mode" to delete either the last point clicked or the last chain drawn.
+3. Double click the final point or select "New Chain" to begin a new chain and repeat for as many chains as you'd like.
+4. When done select the "Plot" button and a popup will appear with a plot of all transects from the first chain and downloading options.
 
-#### Working on a project
+
+### Transect Marker
+* Using this tool multiple "Markers" can be drawn onto a loaded file. These "Markers" can all be saved together as a project and reloaded into the viewer later. When a large file is being explored this allows you to mark out features over multiple sessions.
+
 
 1. Hit the "Transect Marker" button to enter transect mode
 2. Click points along the feature you'd like transects of. Dots will appear on either side of the line drawn indicating the start and end of the transects that will be made
-    * To change the width of the transects being made you may enter the number of pixels into the "Width" text box. Select "Go" to change the width for all future transects.
-3. Select "New Line" to begin a new marker and repeat for as many markers as you'd like
+   * Select "Drag Mode" from the actions sidebar to drag the image without selecting points, and select "Transect Mode" to go back.
+   * Select "Edit Mode" to delete either the last point clicked or the last marker drawn.
+   * To change the width of the transects being made you may enter the number of pixels into the "Width" text box. Select "Set" to change the width for all future transects.
[…]
