### 2024-10-10 0481e66
Revert "Merge branch 'main' into ui_clarity"
This reverts commit b7dbc556cf575c433584c59fe2e9533c4c826a6a, reversing
changes made to 03f01f069345e8658e129df81377805088f230b3.
