+++ tests/test_functions.py
@@ -10,15 +10,11 @@ import unittest
 from PIL import Image as Im
 import numpy as np
 import json
-import pooch
 import xarray as xr
 import nccut.functions as func
 from nccut.multimarker import marker_find
 
-EXAMPLE_JPG_PATH = pooch.retrieve(url="doi:10.5281/zenodo.13883476/example.jpg", known_hash=None)
-EXAMPLE_3D_PATH = pooch.retrieve(url="doi:10.5281/zenodo.13883476/example_3d.nc", known_hash=None)
-EXAMPLE_4V_PATH = pooch.retrieve(url="doi:10.5281/zenodo.13883476/example_4v.nc", known_hash=None)
-PROJECT_EXAMPLE_PATH = pooch.retrieve(url="doi:10.5281/zenodo.13883476/project_example.json", known_hash=None)
+SUPPORT_FILE_PATH = "support/"
 
 
 class Test(unittest.TestCase):
@@ -27,11 +23,11 @@ class Test(unittest.TestCase):
         Test an accurate transect is made when taken horizontally on an image
         """
         # Setup
-        img = Im.open(EXAMPLE_JPG_PATH).convert('RGB')
+        img = Im.open(SUPPORT_FILE_PATH + "example.jpg").convert('RGB')
         points = [1000, 200, 1200, 200]
 
         # App result
-        app = func.ip_get_points(points, img, {"image": EXAMPLE_JPG_PATH})["Cut"]
+        app = func.ip_get_points(points, img, {"image": SUPPORT_FILE_PATH + "example.jpg"})["Cut"]
 
         # Manual result
         arr = np.asarray(img)
@@ -45,11 +41,11 @@ class Test(unittest.TestCase):
         Test an accurate transect is made when taken at 45 on an image
         """
         # Setup
-        img = Im.open(EXAMPLE_JPG_PATH).convert('RGB')
+        img = Im.open(SUPPORT_FILE_PATH + "example.jpg").convert('RGB')
         points = [1000, 200, 1200, 400]
 
         # App result
-        app = func.ip_get_points(points, img, {"image": EXAMPLE_JPG_PATH})["Cut"]
+        app = func.ip_get_points(points, img, {"image": SUPPORT_FILE_PATH + "example.jpg"})["Cut"]
 
         # Manual result
         arr = np.asarray(img)
@@ -66,11 +62,11 @@ class Test(unittest.TestCase):
         Test an accurate transect is made when taken vertically on an image
         """
         # Setup
-        img = Im.open(EXAMPLE_JPG_PATH).convert('RGB')
+        img = Im.open(SUPPORT_FILE_PATH + "example.jpg").convert('RGB')
         points = [1000, 100, 1000, 400]
 
         # App result
-        app = func.ip_get_points(points, img, {"image": EXAMPLE_JPG_PATH})["Cut"]
+        app = func.ip_get_points(points, img, {"image": SUPPORT_FILE_PATH + "example.jpg"})["Cut"]
 
         # Manual result
         arr = np.asarray(img)
@@ -85,7 +81,7 @@ class Test(unittest.TestCase):
         Test an accurate transect is made when taken horizontally on a NetCDF file
         """
         # Setup
-        dat = xr.open_dataset(EXAMPLE_3D_PATH)['Theta'].sel(k=0)
+        dat = xr.open_dataset(SUPPORT_FILE_PATH + "example_3d.nc")['Theta'].sel(k=0)
         config = {"netcdf": {"x": "i", "y": "j", "z": "k", "z_val": "0", "var": "Theta", "file": dat}}
         points = [100, 50, 200, 50]
 
@@ -109,7 +105,7 @@ class Test(unittest.TestCase):
[…]
