### 2024-10-10 76ef983
Revert "Merge branch 'main' into ui_clarity"
This reverts commit d5dea22eb056de141ee190e5a8da695b77326a20, reversing
changes made to 17d0ebbc3b4a1fb722d5fceb9f22669059299322.
