+++ docs/source/example.rst
@@ -1,7 +1,7 @@
 Tutorials
 =========
 
-To test out NcCut you can download some example files from here: `https://zenodo.org/records/13883476 <https://zenodo.org/records/13883476>`_:
+To test out NcCut you can download some example files from the **NcCut** `repository <https://github.com/rchartra/NcCut/tree/master/support>`_:
 
 * ``example.jpg`` is a SAR image of sea ice floes in the arctic ocean from the Sentinel-1 satellite
 * ``project_example.json`` is an example project file output from ``example_4v.nc`` to practice uploading projects
