+++ docs/source/example.rst
@@ -1,7 +1,8 @@
 Tutorials
 =========
 
-To test out NcCut you can download some example files from the **NcCut** `repository <https://github.com/rchartra/NcCut/tree/master/support>`_:
+To test out NcCut you can download some example files from here: `https://zenodo.org/records/13883476 <https://zenodo.org/records/13883476>`_:
+
 
 * ``example.jpg`` is a SAR image of sea ice floes in the arctic ocean from the Sentinel-1 satellite
 * ``project_example.json`` is an example project file output from ``example_4v.nc`` to practice uploading projects
@@ -9,12 +10,6 @@ To test out NcCut you can download some example files from the **NcCut** `reposi
 * ``example_4v.nc`` is a dataset of various ocean circulation metrics calculated from velocity data from the MITgcm LLC4320 Model near the gulf stream
 * ``test_img.png`` is simply an image of a square to use to validate the transect tools' accuracy
 * ``test_project_file.json`` is another example project file for practicing uploading project files
-
-Using NcCut with a NetCDF File
----------------------------------
-
-.. video:: _media/NcCut_Tutorial.mp4
-    :width: 650
 
 .. _output:
 
