### 2024-10-10 rchartra/NcCut#24 @rchartra
Okay I've cleaned out large files from my commit history and the repository is now at ~18mb
