### 2024-10-10 rchartra/NcCut#28 @rchartra
I was able to replace all cv2 elements with Matplotlib, so OpenCV is no longer a dependency
