### 2024-10-10 rchartra/NcCut#27 @rchartra
After testing it appears Cython is not required for Kivy to install properly
