### 2024-10-10 rchartra/NcCut#22 @rchartra
This has been fixed in v0.4.6
