### 2024-10-10 rchartra/NcCut#21 @rchartra
This has been fixed (including tests) in v0.4.6
