### 2024-10-10 rchartra/NcCut#23 @rchartra
These issues have all been resolved in v0.4.6
