### 2024-10-11 rchartra/NcCut#16 @rchartra
Alright so I've made some changes to the UI to address this issue:

- The sidebar updates with each step so no buttons are ever shown other than the ones you can use at the current point in the workflow
- You can double click to end a chain/marker in addition to the "New <tool>" button on the sidebar
- Button names have been changed to be more clear
- Static UI elements and buttons are more clearly distinct

I'm leaving this issue open for now until I get more opinions from my colleagues on these updates.
