### 2024-10-16 c3ea7e4
Save global and netcdf specific metadata with output
