### 2024-10-27 rchartra/NcCut#31 @Huite
Windows: installation instructions mentions `python3`, but Windows generally uses `python`
Related to: https://github.com/openjournals/joss-reviews/issues/7185

(Nice work on the fixes! -- I'm having another look today)

On my Windows machine, I have to run `python` instead of `python3`; it won't recognize python3.

As far I know, since Linux relied on a system python (and with the issues of the 2-to-3 transition), `python` historically points to a python 2.

I think if Python is installed through the microsoft store, `python3` will also work, next to to `python`. I mostly work with (mini)conda and pixi, and on Windows, it makes `python` available but not `python3`.

So my recommendation is to use `python` in the Windows installation instructions, that should work in the majority of cases.
