### 2024-10-27 rchartra/NcCut#32 @Huite
Add a `__main__.py`
Related to: https://github.com/openjournals/joss-reviews/issues/7185

If you add a `__main__.py` to your module:

```python
from nccut.nccut import NcCut

if __name__ == "__main__":
    NcCut().run()
```

This will enable running `python -m nccut` as well, since that command looks for for a main file inside your module.

Secondly, I see poetry creates a `nccut.exe` :

```toml
[tool.poetry.scripts]
nccut = "nccut.nccut:run"
```

My recommendation is to put the run function into the main directory, since it's generally considered a Python best practice to define your main entry point in a `__main__.py` file. Then the poetry command becomes `nccut = "nccut.__main__:run"`.

Then your main file would basically look like:

```python
import argparse
import nccut

def run():
     ...

if __name__ == "__main__":
     run()
```

It separates the entry point logic from your core functionality, but most importantly it's a clear, standard location for entry points that other developers will expect.
