### 2024-10-27 rchartra/NcCut#33 @Huite
Variables are unnecessarily abbreviated in Plotting Menu dropdown menu
Related to: https://github.com/openjournals/joss-reviews/issues/7185

See here:
![image](https://github.com/user-attachments/assets/df33d2e1-541c-4b96-9973-24bb537753e3)

Despite the variable name `Rosby Number` being relatively short and and an abundace of screen space, it is abbreviated.
Variables often share some prefix, so this makes identifying the right variable quite challenging in some cases.
