### 2024-10-27 rchartra/NcCut#35 @Huite
X and Y coordinates are assumed to be equidistant
Related to: https://github.com/openjournals/joss-reviews/issues/7185

Skipping coordinates makes life a lot easier, but it also means that you have to assume the grid data is equidistant.

### Resampling to an equidistant rectilinear grid
If this isn't the case, you basically need to resample the data so the plot is representative.

For axis aligned coordinates, resampling to an equidistant grid is easy with numpy searchsorted, but that isn't the case for curvilinear grids. And that opens a whole can of worms of having to deal with coordinates.

### Make a note
Alternatively, you could add a note in your documentation about the assumptions NcCut makes.
