### 2024-10-27 rchartra/NcCut#37 @Huite
No way to cancel drawing a transect chain
Related to: https://github.com/openjournals/joss-reviews/issues/7185

When drawing a transect, there seems to be no way of cancelling. Once you start drawing, you can only complete it by double-clicking.
No distinction is made between the left and right mouse button (LMB / RMB): both add points, and double clicking "commits" the chain. Pressing escape does not cancel it.

I'm personally used to QGIS, and its linestring vector editing works as follows:

1. Press LMB to add a vertex
2. Press RMB to "commit"
3. Press Escape to cancel the entire chain

I would avoid double-clicking for precision tasks, especially when you have good other options available. The risk with double clicking is that you shift the mouse between clicks, thereby making it two clicks instead of a double click. (Operating systems even need a setting to control what counts as a double click because if a user has motor control problems, double clicking is much more challenging than other mouse operations.)
