### 2024-10-27 rchartra/NcCut#38 @Huite
Quit button closes immediately
Related to: https://github.com/openjournals/joss-reviews/issues/7185

If you click the Quit Button it closes immediately. In most programs, quitting via the in-application menu is two or three steps:

* Click File Menu (or something similar)
* Click Close
* If there's unsaved work, it will now still show a confirmation dialog)

In NcCut, the Quit button is right there on screen, and it closes the app immediately. This means there's no room for a misclick.
I'd suggest opening a confirmation dialog giving the user the option to cancel in case of a misclick.
