### 2024-10-27 rchartra/NcCut#39 @Huite
Transect Marker: symbology is difficult to visually parse
Related to: https://github.com/openjournals/joss-reviews/issues/7185

See this somewhat contrived example:

![image](https://github.com/user-attachments/assets/61c92b1b-92fe-415a-9644-5aa09788a9f8)

If I pay attention, I can deduce the orthogonal transects... but wouldn't this be easier?

![image](https://github.com/user-attachments/assets/3813a691-5cdb-438a-bf10-090a943acb46)

I think the optimal solution would be to make the chain dashed, and the orthogonal transects solid. This makes it consistent with the Transect Chain symbology, where the transects that are plotted are also solid.
