### 2024-10-27 rchartra/NcCut#40 @Huite
Transect Marker: name is a bit confusing
Related to: https://github.com/openjournals/joss-reviews/issues/7185

I find the name transect marker a bit confusing. Your documentation mentions:

> The [Transect Marker](https://nccut.readthedocs.io/en/latest/usage.html#transect-marker) tool allows you to draw a line along a feature, and transects will be automatically made orthogonal to the line you’ve drawn with a set width:

It seems like the orthogonal part is the most important here. "Orthogonal Transect Chain" or "Perpendicular Transect Chain" is a bit of a mouthful, but I feel like either "Orthogonal Chain" or "Orthogonal Transects" cover the functionality better than "Marker", which is very general. At the very least, I think "orthogonal" or "perpendicular" should be part of the name.
