### 2024-10-27 rchartra/NcCut#41 @Huite
Transect Marker: width setting
Related to: https://github.com/openjournals/joss-reviews/issues/7185

A few things:

### Ambiguous unit
This is may be due to my own preconceptions, but I was confused as to what the units of the width field are. The documentation mentions this, but a label mentioning pixels makes it immediately clear.

### Implicit default
By default, the width setting is empty. This makes it unclear what the default value is. It should simply show the default value -- I know it exists, because even without filling a value, it does generate the markers.

### Lack of validation
I can type an arbitrary string ("abcdef") into the width window. It should validate that the input is a valid integer when "set" is clicked and raise a warning or something when the input isn't valid.

I can currently type in floats too: does that mean I'll get half a pixel?

### Hidden limits
I see this in the code:

```python
        self.min = 1
        self.max = 400
```

I.e. there's only room between 1 and 400 pixels. Having a lower limit of 1 pixel is obviously reasonable, but an upper bound of 400 pixels isn't necessary. Rather I would make the upper limit equal to something like `max(nrow, ncol)`: after all, why shouldn't I be able to make transects that cut through my entire dataset?

### Different widget
In Qt, I would use a SpinBox, since the input is discrete (1 pixel increments), it clearly communicates that it is discrete. The beauty of a spin box is also that you can type a value into it, but it'll only accept valid values.
It seems like kivy doesn't have an out-of-the-box solution. You could also think of a slider with a synchronized textbox. The textbox would automatically convert anything numeric to an int. The slider clearly communicates the upper and lower boundary.

### Visual aid
There are no axis ticks or anything, and pixels aren't necessarily the easiest unit to gauge: essentially, you're asking the user to keep in mind the total bounding box of their data, identify the scale of interest (e.g. length of Florida), divide that by their bounding box size, and multiply times the total number of pixels -- or try to it via trial and error.

You can either show axis ticks or scale bar. Axis ticks seem easier at first glance. After all, what should the size be of a scale bar be? (100 pixels? More? Less?)
