### 2024-10-30 5c7f7d6
Add configuration file for setting defaults and adding additional metadata
