+++ tests/test_gui.py
@@ -382,6 +382,11 @@ class Test(unittest.TestCase):
         """
         Test netcdf configuration popup ensures valid netcdf configuration settings
         """
+        run_app.home.general_config = {"graphics_defaults": {"contrast": 0, "line_color": "Blue", "colormap": "viridis",
+                                                             "circle_size": 5},
+                                       "netcdf": {"dimension_order": ["z", "y", "x"]},
+                                       "tool_defaults": {"marker_width": 40},
+                                       "metadata": {}}
         run_app.home.ids.file_in.text = TEST_NC_PATH
         run_app.home.load_btn()
         popup = run_app.home.nc_popup
@@ -570,9 +575,6 @@ class Test(unittest.TestCase):
             self.assertListEqual(list(res1["Vorticity"].keys()), ["Marker 1", "Vorticity_attrs"],
                                  "Only selected marker and metadata should be saved")
             self.assertIn("global_metadata", list(res1.keys()), "Global metadata not added")
-            self.assertListEqual(list(res1["global_metadata"].keys()),
-                                 ["time_stamp", "user", "license", "file", "netcdf_attrs", "dim_attrs"],
-                                 "Not all metadata fields present")
             self.assertEqual(len(list(res1["Vorticity"]["Marker 1"].keys())), 6,
                              "Marker should have 3 transects and Click x, Click y, Width")
         # Plot Saving
@@ -603,9 +605,6 @@ class Test(unittest.TestCase):
             self.assertListEqual(list(res2["Vorticity"].keys()), ["Marker 1", "Marker 2", "Vorticity_attrs"],
                                  "Two markers were selected so two should be saved")
             self.assertIn("global_metadata", list(res2.keys()), "Global metadata not added")
-            self.assertListEqual(list(res2["global_metadata"].keys()),
-                                 ["time_stamp", "user", "license", "file", "netcdf_attrs", "dim_attrs"],
-                                 "Not all metadata fields present")
             self.assertEqual(len(list(res2["Vorticity"]["Marker 1"].keys())), 4,
                              "Marker 1 should have 1 transect and Click x, Click y, Width")
             self.assertEqual(list(res2["Vorticity"]["Marker 1"].keys())[0], "Cut 1",
@@ -688,3 +687,65 @@ class Test(unittest.TestCase):
         plot_popup.get_all_z_plot()
         self.assertNotEqual(og_plot, str(plot_popup.plot),
                             "If only one transect is selected the all z plot should be created")
+
+    def test_metadata_and_config_file(self):
+        """
+        Tests whether a passed valid configuration dictionary is properly used to change settings in the app.
+        """
+        non_default_config = {"graphics_defaults": {"contrast": 6, "line_color": "Green", "colormap": "plasma",
+                                                    "circle_size": 50},
+                              "netcdf": {"dimension_order": ["z", "x", "y"]},
+                              "tool_defaults": {"marker_width": 20},
+                              "metadata": {"new_field": "new_val"}}
+        run_app.home.general_config = non_default_config
+        run_app.home.ids.file_in.text = EXAMPLE_3D_PATH
+        run_app.home.load_btn()
+        popup = run_app.home.nc_popup
+        self.assertEqual(popup.x_select.text, "j", "Dimension order was not updated from config file.")
+        self.assertEqual(popup.y_select.text, "i", "Dimension order was not updated from config file.")
+        self.assertEqual(popup.z_select.text, "k", "Dimension order was not updated from config file.")
+        self.assertEqual(popup.depth_select.text, "0", "Dimension order was not updated from config file.")
+        popup.load.dispatch("on_press")
+        popup.load.dispatch("on_release")
+
[…]
