### 2024-10-30 336ac89
Determine if cython resolves macos error
