+++ src/nccut/plotpopup.py
@@ -329,33 +329,31 @@ class PlotPopup(Popup):
         Args:
             f_path (str): Output file path
         """
-        print("method called")
         if f_path.find(".") == -1:
             f_path = f_path + ".json"
         else:
             f_path = f_path[:f_path.find(".")] + ".json"
-        # try:
-        dat = copy.deepcopy(self.active_data)
-        if len(self.active_vars) == 0:  # If Image
-            final = self.add_group_info(dat)
-        elif len(self.active_z) == 0:  # If 2D NetCDF
-            final = {}
-            for var in list(dat.keys()):
-                final[var] = self.add_group_info(dat[var])
-        else:  # If 3D NetCDF
-            final = {}
-            for var in list(dat.keys()):
-                final[var] = {}
-                for z in list(dat[var].keys()):
-                    final[var][z] = self.add_group_info(dat[var][z])
-        final = self.add_metadata(final)
-        print("_FILE_PATH_SAVED_: " + f_path)
-        with open(f_path, "w") as f:
-            json.dump(final, f)
+        try:
+            dat = copy.deepcopy(self.active_data)
+            if len(self.active_vars) == 0:  # If Image
+                final = self.add_group_info(dat)
+            elif len(self.active_z) == 0:  # If 2D NetCDF
+                final = {}
+                for var in list(dat.keys()):
+                    final[var] = self.add_group_info(dat[var])
+            else:  # If 3D NetCDF
+                final = {}
+                for var in list(dat.keys()):
+                    final[var] = {}
+                    for z in list(dat[var].keys()):
+                        final[var][z] = self.add_group_info(dat[var][z])
+            final = self.add_metadata(final)
+            with open(f_path, "w") as f:
+                json.dump(final, f)
 
-        func.alert_popup("Download Complete")
-        # except Exception as error:
-        # func.alert_popup(str(error))
+            func.alert_popup("Download Complete")
+        except Exception as error:
+            func.alert_popup(str(error))
 
     def download_all_data(self, f_path):
         """
