+++ .github/workflows/python-app.yml
@@ -76,7 +76,7 @@ jobs:
       - name: Set up Python 3.x
         uses: actions/setup-python@v5
         with:
-          python-version: ${{ matrix.version }}
+          python-version: '3.11'
       - name: Install Poetry
         run: |
           python -m pip install --upgrade pip
