+++ pyproject.toml
@@ -22,7 +22,6 @@ scipy = "^1.13.1"
 matplotlib = "^3.9.0"
 xarray = "^2024.6.0"
 netCDF4 = "^1.7.1" # Necessary for xarray
-cython = "^3.0.11"
 Kivy = "^2.3.0"
 progress = "^1.6"
 plyer = "^2.1.0"
