### 2024-10-30 1385611
Debug config workflow
