### 2024-10-30 22a5ac8
Debug config workflow
