### 2024-10-30 21b14c7
Debug config workflow
