+++ src/nccut/functions.py
@@ -139,8 +139,10 @@ def find_config(config_file):
     config_path = os.getenv('NCCUT_CONFIG')
     if config_file:
         config_path = config_file
+        print("Config found via argument")
     elif config_path:
         config_path = Path(config_path)
+        print("Config found via environment variable")
     else:
         # Look for file in working directory
         config_path = Path.cwd() / "nccut_config.toml"
@@ -148,8 +150,12 @@ def find_config(config_file):
             # Look for file in default location
             if os.name == 'nt':  # Windows
                 config_path = Path(os.getenv('APPDATA')) / 'nccut' / 'nccut_config.toml'
+                print("Config found in default location: windows")
             else:  # Unix-based systems (Linux/macOS)
                 config_path = Path.home() / ".config" / "nccut" / "nccut_config.toml"
+                print("Config found in default location: linux")
+        else:
+            print("Config found in working directory")
     try:
         with open(config_path, 'rb') as config_file:
             config = tomli.load(config_file)
