### 2024-10-30 2f48d4c
Debug config workflow
