+++ .github/workflows/python-app.yml
@@ -83,12 +83,14 @@ jobs:
         python -m poetry run python -c 'from nccut.functions import find_config; assert find_config("") == {"graphics_defaults": {"contrast": 0}}, "Config load failed via working directory"'
     - name: Test with Default Path
       run: |
+        echo %APPDATA%
         mkdir %APPDATA%\nccut
         move nccut_config.toml %APPDATA%\nccut\nccut_config.toml
-        python -m poetry run python -c 'from nccut.functions import find_config; assert find_config("") == {"graphics_defaults": {"contrast": 0}}, "Config load failed via working directory"'
+        dir %APPDATA%\nccut
+        python -m poetry run python -c 'from nccut.functions import find_config; assert find_config("") == {"graphics_defaults": {"contrast": 0}}, "Config load failed via default path"'
     - name: Test with command line argument
       run: |
-        python -m poetry run python -c 'from nccut.functions import find_config; assert find_config("nccut_config.toml") == {"graphics_defaults": {"contrast": 0}}, "Config load failed via working directory"'
+        python -m poetry run python -c 'from nccut.functions import find_config; assert find_config("nccut_config.toml") == {"graphics_defaults": {"contrast": 0}}, "Config load failed via command line"'
 
   mac_deps_test:
     runs-on: macos-latest
