### 2024-10-30 c633264
Debug config workflow
