### 2024-10-30 9f62c19
Debug config workflow
