+++ .github/workflows/python-app.yml
@@ -81,12 +81,15 @@ jobs:
     - name: Test Config in Working Directory
       run: |
         python -m poetry run python -c 'from nccut.functions import find_config; assert find_config("") == {"graphics_defaults": {"contrast": 0}}, "Config load failed via working directory"'
+    - name: Set APPDATA
+      run: |
+        set APPDATA=C:\Users\runneradmin\AppData\Roaming
     - name: Test with Default Path
       run: |
         echo "APPDATA location is ${{ env.APPDATA }}"  # Confirm the APPDATA path
-        mkdir "${{ env.APPDATA }}\nccut"
-        move nccut_config.toml "${{ env.APPDATA }}\nccut\nccut_config.toml"
-        dir "${{ env.APPDATA }}\nccut"  # Verify that the file moved to the correct location
+        mkdir "%APPDATA%\nccut"
+        move nccut_config.toml "%APPDATA%\nccut\nccut_config.toml"
+        dir "%APPDATA%\nccut"  # Verify that the file moved to the correct location
         python -m poetry run python -c "from nccut.functions import find_config; assert find_config('') == {'graphics_defaults': {'contrast': 0}}, 'Config load failed via default path'"    - name: Test with command line argument
     - name: Test with command line argument
       run: |
