### 2024-10-30 e35d6da
Debug config workflow
