### 2024-10-30 19bb5f9
Debug config workflow
