### 2024-10-30 a562eb3
Debug config workflow
