### 2024-10-30 9edc512
Debug config workflow
