### 2024-10-30 7ae0f40
Debug config workflow
