+++ .github/workflows/python-app.yml
@@ -94,7 +94,7 @@ jobs:
         mkdir command_dir
         copy nccut_config.toml command_dir\nccut_config.toml
         dir command_dir
-        python -m poetry run python -c 'from nccut.functions import find_config; assert find_config("command_dir\nccut_config.toml") == {"graphics_defaults": {"contrast": 0}}, "Config load failed via command line"'
+        python -m poetry run python -c 'from nccut.functions import find_config; assert find_config("D:\a\NcCut\NcCut\command_dir") == {"graphics_defaults": {"contrast": 0}}, "Config load failed via command line"'
 
   mac_deps_test:
     runs-on: macos-latest
