### 2024-10-30 a0a6aee
Debug config workflow
