+++ .github/workflows/python-app.yml
@@ -93,8 +93,9 @@ jobs:
       run: |
         mkdir command_dir
         copy nccut_config.toml command_dir\nccut_config.toml
-        dir command_dir
-        python -m poetry run python -c 'from nccut.functions import find_config; assert find_config("D:\a\NcCut\NcCut\command_dir") == {"graphics_defaults": {"contrast": 0}}, "Config load failed via command line"'
+        python -m poetry run nccut -config custom_dir\nccut_config.toml > output.log
+        type output.log
+        findstr /c:"Valid configuration file found at custom_dir\\nccut_config.toml" output.log
 
   mac_deps_test:
     runs-on: macos-latest
