### 2024-10-30 1f679af
Debug config workflow
