### 2024-10-30 7e057cb
Debug config workflow
