### 2024-10-30 49fe6e4
Debug config workflow
