+++ .github/workflows/python-app.yml
@@ -91,11 +91,11 @@ jobs:
         python -m poetry run python -c "from nccut.functions import find_config; assert find_config('') == {'graphics_defaults': {'contrast': 0}}, 'Config load failed via default path'"    - name: Test with command line argument
     - name: Test with command line argument
       run: |
-        mkdir ${GITHUB_WORKSPACE}\command_dir
-        copy nccut_config.toml ${GITHUB_WORKSPACE}\command_dir\nccut_config.toml
-        python -m poetry run nccut -config command_dir\nccut_config.toml > output.log
+        mkdir "${{ github.workspace }}\command_dir"
+        copy nccut_config.toml "${{ github.workspace }}\command_dir\nccut_config.toml"
+        python -m poetry run nccut -config "${{ github.workspace }}\command_dir\nccut_config.toml" > output.log
         type output.log
-        findstr /c:"Valid configuration file found at command_dir\\nccut_config.toml" output.log
+        findstr /c:"Valid configuration file found at ${{ github.workspace }}\\nccut_config.toml" output.log
 
   mac_deps_test:
     runs-on: macos-latest
