### 2024-10-30 eb505f9
Debug config workflow
