### 2024-10-31 807cb0f
Debug config workflow
