+++ .github/workflows/python-app.yml
@@ -106,12 +106,12 @@ jobs:
         mkdir "${{ env.APPDATA_DIR }}\nccut"
         copy nccut_config.toml "${{ env.APPDATA_DIR }}\nccut\nccut_config.toml"
         dir "${{ env.APPDATA_DIR }}\nccut"  # Verify that the file moved to the correct location
-        python -m poetry run python -c "from nccut.functions import find_config; assert find_config('') == {'graphics_defaults': {'contrast': 0}}, 'Config load failed via default path'"    - name: Test with command line argument
+        python -m poetry run python -c "from nccut.functions import find_config; assert find_config('') == {'graphics_defaults': {'contrast': 0}}, 'Config load failed via default path'"
     - name: Test with command line argument
       run: |
         mkdir "${{ github.workspace }}\command_dir"
         copy nccut_config.toml "${{ github.workspace }}\command_dir\nccut_config.toml"
-        python -m poetry run python -c "from nccut.functions import find_config; assert find_config("${{ github.workspace }}\command_dir\nccut_config.toml") == {'graphics_defaults': {'contrast': 0}}, 'Config load failed via default path'"    - name: Test with command line argument
+        python -m poetry run python -c "from nccut.functions import find_config; assert find_config("${{ github.workspace }}\command_dir\nccut_config.toml") == {'graphics_defaults': {'contrast': 0}}, 'Config load failed via default path'"
 
   mac_deps_test:
     runs-on: macos-latest
