### 2024-10-31 44726b7
Debug config workflow
