### 2024-10-31 c72c318
Debug config workflow
