+++ .github/workflows/python-app.yml
@@ -153,7 +153,7 @@ jobs:
     - name: Test with Default Path
       run: |
         mkdir -p "${{ env.CONFIG_DIR }}/nccut"
-        cp nccut_config.toml "${{ env.APPDATA_DIR }}/nccut"
+        cp nccut_config.toml "${{ env.CONFIG_DIR }}/nccut"
         ls "${{ env.CONFIG_DIR }}/nccut"  # Verify that the file moved to the correct location
         python -m poetry run python -c "from nccut.functions import find_config; assert find_config('') == {'graphics_defaults': {'contrast': 0}}, 'Config load failed via default path'"
     - name: Test with command line argument
