+++ .github/workflows/python-app.yml
@@ -159,7 +159,7 @@ jobs:
     - name: Test with command line argument
       run: |
         mkdir -p "${{ github.workspace }}/command_dir"
-        copy nccut_config.toml "${{ github.workspace }}/command_dir"
+        cp nccut_config.toml "${{ github.workspace }}/command_dir"
         python -m poetry run python -c 'from nccut.functions import find_config; assert find_config(r"${{ github.workspace }}/command_dir/nccut_config.toml") == {"graphics_defaults": {"contrast": 0}}, "Config load failed via command line argument path"'
 
   mac_deps_test:
