+++ src/nccut/project_file_printer.py
@@ -10,10 +10,11 @@ def project_file_printer(data, tab):
         data (dict): Dictionary output from NcCut
         tab (int): Always use 0, used in the recursive process for indentation
     """
-    for key in list(data.keys()):
-        print(" " * tab, end="")
-        if isinstance(data[key], list):
-            print(key + ": [" + str(data[key][0]) + ", " + str(data[key][1]) + ", ..." + "]")
-        else:
-            print(key + ":")
-            project_file_printer(data[key], tab + 4)
+    if isinstance(data, dict):
+        for key in list(data.keys()):
+            print(" " * tab, end="")
+            if isinstance(data[key], list):
+                print(key + ": [" + str(data[key][0]) + ", " + str(data[key][1]) + ", ..." + "]")
+            else:
+                print(key + ":")
+                project_file_printer(data[key], tab + 4)
