### 2024-11-06 rchartra/NcCut#31 @rchartra
Updated in #42
