### 2024-11-06 rchartra/NcCut#32 @rchartra
Thanks for the tip! I've implemented this in #42
