### 2024-11-06 rchartra/NcCut#38 @platipodium
I think I disagree with @Huite on the solution, but not on the problem analysis.   The sudden/unexpected closing of the app is very well in line with ncview's behavior :=(. On the other hand, it is a quicklook tool, and as such should be quick to dismiss, as well.  I would feel obstructed by a "do you really want to close?" dialog.

If there is no data loss on quick-close (e.g. by auto-saving the view) I'd have no problem with the one-click close functionality.
