### 2024-11-06 rchartra/NcCut#36 @rchartra
Good point, I agree that in both the main screen and the plotting popup menu as much space as possible should be given to the plot. Kivy is also kind of finicky with it's size hinting but once you figure out the pattern it's not too bad. So I'll figure out a way to make the layout better. I also like the idea of making the font size configurable. I'll add it as an option for the nccut_config.toml file.
