### 2024-11-06 rchartra/NcCut#35 @rchartra
For the interactive plot in the plot menu I plot on a non equidistant grid according to the coordinate values so I can also implement that for the main interactive plot.
