### 2024-11-06 rchartra/NcCut#37 @rchartra
Currently, you can cancel drawing a chain by going Edit Mode -> Delete Last Line/Point, though I agree that having shortcuts for performing common actions like that would be nice. The QGIS editing shortcuts seem very reasonable so I'll implement those.
