### 2024-11-08 rchartra/NcCut#38 @rchartra
One possibility is to show the confirmation dialog, but have an option to select "Don't show this again" so the user can choose to one-click quit on future occasions.
