### 2024-11-08 rchartra/NcCut#38 @Huite
I'm often somewhat annoyed by the "Are you sure?" prompts too... and this is a quicklook tool indeed.

Some thoughts:

* I think I opened this issue specifically because I misclicked, which happens easily because the Quit button is near the other buttons.
* The application can be also closed directly via either the taskbar, or via the `X` in the window corner. These are much less likely to be hit accidentally. I use several application which don't even have an in-GUI quit button.
* A "File" menu with a Quit option is already two clicks, and thereby guards against misclicks, but that's not a good fit for NcCut.
* The "Don't show this again" sounds reasonable, but you'll have to start storing user settings and for most applications that do this, I wouldn't know how to re-enable showing the dialog.
