### 2024-11-20 dad860b
UI More resilient to resizing. Font can be adjusted in config file
