+++ src/nccut/filedisplay.py
@@ -13,6 +13,7 @@ ScatterLayout class. Manages the creation and deletion of tools.
 import kivy
 from kivy.graphics.transformation import Matrix
 from kivy.uix.scatterlayout import ScatterLayout
+from kivy.core.window import Window
 import kivy.uix as ui
 from kivy.core.image import Image as CoreImage
 from kivy.metrics import dp
@@ -59,6 +60,7 @@ class FileDisplay(ScatterLayout):
         cir_size (float): Circle size for transect tools
         cmaps (dict): Dictionary of colormap names mapping to colormap values from cv2
         colormap: current colormap data
+        btn_height: Height for buttons in sidebar which adapts to font size
         transect_chain_btn: Transect Chain button which opens Transect Chain tool
         transect_marker_btn: Transect Marker button which opens Transect Marker tool
         initial_side_bar (list): List of sidebar buttons in 'Tools' sidebar menu
@@ -112,56 +114,62 @@ class FileDisplay(ScatterLayout):
 
         self.cmaps = plt.colormaps()[:87]
         self.colormap = g_config["colormap"]
-
+        self.btn_height = dp(20) + self.home.font
         # Initial Sidebar Widgets
-        self.transect_chain_btn = func.RoundedButton(text="Transect Chain", size_hint=(1, 0.1),
+        self.transect_chain_btn = func.RoundedButton(text="Transect Chain", size_hint_y=None, height=self.btn_height,
                                                      font_size=self.home.font)
         self.transect_chain_btn.bind(on_press=lambda x: self.transect_btn("transect_chain"))
 
-        self.transect_marker_btn = func.RoundedButton(text="Transect Marker", size_hint=(1, 0.1),
+        self.transect_marker_btn = func.RoundedButton(text="Transect Marker", size_hint_y=None, height=self.btn_height,
                                                       font_size=self.home.font)
         self.transect_marker_btn.bind(on_press=lambda x: self.transect_btn("transect_marker"))
         self.initial_side_bar = [self.transect_chain_btn, self.transect_marker_btn]
 
         # Action Widgets
-        self.drag_btn = func.RoundedButton(text="Drag Mode", size_hint=(1, 0.1),
+        self.drag_btn = func.RoundedButton(text="Drag Mode", size_hint_y=None, height=self.btn_height,
                                            halign='center', valign='center', font_size=self.home.font)
         self.drag_btn.bind(on_press=self.drag_mode)
-        self.edit_btn = func.RoundedButton(text="Edit Mode", size_hint=(1, 0.1),
+        self.edit_btn = func.RoundedButton(text="Edit Mode", size_hint_y=None, height=self.btn_height,
                                            halign='center', valign='center', font_size=self.home.font)
         self.edit_btn.bind(on_press=self.edit_mode)
-        self.close_tool_btn = func.RoundedButton(text="Close Tool", size_hint=(1, 0.1),
+        self.close_tool_btn = func.RoundedButton(text="Close Tool", size_hint_y=None, height=self.btn_height,
                                                  halign='center', valign='center', font_size=self.home.font)
         self.close_tool_btn.bind(on_press=self.close_tool)
         self.action_widgets = [self.close_tool_btn, self.drag_btn, self.edit_btn]
         self.tool_action_widgets = self.action_widgets
         # Drag Mode Widgets
-        self.tran_mode_btn = func.RoundedButton(text="Transect Mode", size_hint=(1, 0.1),
+        self.tran_mode_btn = func.RoundedButton(text="Transect Mode", size_hint_y=None, height=self.btn_height,
                                                 halign='center', valign='center', font_size=self.home.font)
         self.tran_mode_btn.bind(on_press=self.drag_mode)
 
[…]
