### 2024-11-20 rchartra/NcCut#36 @rchartra
Resolved in v0.4.7
