### 2024-11-25 #7185 @hugoledoux
@rchartra can you please update us on the status of the opened issues?
