### 2024-11-25 rchartra/NcCut#39 @rchartra
This is a good point and I like this idea. I'll implement the solid lines for the orthogonal transects and the dashed lines between the clicked points. Kivy shockingly does not have a way to make dashed lines if the width of the line is greater than one, so I need to be a little creative to make this to work.
