### 2024-11-25 rchartra/NcCut#41 @rchartra
You make a good point that "pixels" is not very clear or relevant to a user. I think the ideal thing to do is to change the unit to come from the coordinates of the NetCDF file (if they make sense and are available) and revert back to pixels otherwise (like with JPEGs).

A visual aid seems like a generally good idea for this purpose but also just overall for the user to have a sense of the scale of the dataset. Axis ticks are the most clear though implementing them in a way that scales as the user zooms, moves as the user drags the image around, and maintains "nice" numbers throughout is quite tricky. I have done it before with the plotting menu though. A scale bar would be easier I think as it would only have to adapt to the scale of the image as the user zooms in (I would do it with coordinate values not pixels). I'm still deciding between the two and either way I will wait until I have implemented non-equidistant coordinates so the scaling is accurate.

As for the width adjustment widget itself, Kivy allows you to control a lot about text entry boxes. I can probably mimic a lot of the validating behaviors of a spinbox myself. To indicate maximum values I'm thinking of turning the widget adjustment into a popup or a dropdown menu which will allow more room to add that information as a label.
