### 2024-11-25 rchartra/NcCut#43 @rchartra
I've looked into this issue and it seems to be a bug with the file selection widget from Kivy on MacOS (see [here](https://github.com/kivy/pyobjus/issues/21)). It seems it's been fixed but not released and I don't think they are planning on doing a release any time soon unfortunately. The FileChooser works fine on the Macs I have access to so I'm hoping I can figure out what the difference is. Otherwise there's other python UI frameworks that provide file selection functionality but I was running into issues with those before as mixing frameworks is generally not well supported.

For now I will add an error popup so at least the app won't crash and the user can still type in the file path manually.
