### 2024-11-25 #7185 @rchartra
I am currently in the process of implementing non equidistant coordinates to resolve [NcCut#35](https://github.com/rchartra/NcCut/issues/35).
