### 2024-11-25 rchartra/NcCut#43 @platipodium
I am happy with a temporary fix; that way, I can continue testing your software.  For now, this is a showstopper also for my review, I'm afraid.
