### 2024-11-25 rchartra/NcCut#43 @rchartra
Sounds good, I'll let you know when it's in place.
