### 2024-12-13 rchartra/NcCut#44 @platipodium
pip dependency conflict: fortls 3.1.2 requires json5, which is not installed.
With the instructions ` pip install nccut` I get the following error

```
Installing collected packages: pytz, pyobjus, progress, plyer, urllib3, tzdata, tomli, six, pyparsing, pygments, pillow, packaging, numpy, kiwisolver, idna, fonttools, docutils, cycler, charset-normalizer, certifi, scipy, requests, python-dateutil, contourpy, cftime, pandas, netCDF4, matplotlib, Kivy-Garden, xarray, Kivy, nccut
ERROR: pip's dependency resolver does not currently take into account all the packages that are installed. This behaviour is the source of the following dependency conflicts.
fortls 3.1.2 requires json5, which is not installed.
```

This is fixed by doing `pip install json5`, so it seems to be missing from dependencies.
