### 2024-12-13 #7185 @platipodium
New issue:
 - [ ] https://github.com/rchartra/NcCut/issues/44
