### 2024-12-13 rchartra/NcCut#45 @platipodium
When resizing window, plot does not grow
part of my review at https://github.com/openjournals/joss-reviews/issues/7185

After loading a picture, when I resize the window, the map does not adjust to the new size of the window.
