### 2024-12-13 rchartra/NcCut#46 @platipodium
Figure does not grow/shrink with window
part of my review at https://github.com/openjournals/joss-reviews/issues/7185

After loading a picture, when I resize the window, the map does not adjust to the new size of the window.
