### 2024-12-13 #7185 @platipodium
And another one
- [ ] https://github.com/rchartra/NcCut/issues/46
