### 2024-12-17 a129edb
- Data with nonuniform coordinates is interpolated to match coordinate data
