+++ src/nccut/functions.py
@@ -21,11 +21,11 @@ import kivy.uix as ui
 from functools import partial
 from scipy.interpolate import RegularGridInterpolator, CubicSpline
 import numpy as np
+from PIL import Image as im
 import math
 import io
 import warnings
 import itertools
-import xarray as xr
 import matplotlib
 from pathlib import Path
 import tomli
@@ -507,72 +507,82 @@ def get_color_bar(colormap, data, face_color, text_color, font):
     return plot
 
 
-def sel_data(config):
+def subset_around_transect(config, points):
     """
-    Selects data from NetCDF file according to dimension and variable selections.
+    Determines and loads a subset of the data that surrounds the transect.
 
     Args:
-        config: Dictionary of key details about NetCDF file as outlined in
-            :meth:`nccut.netcdfconfig.NetCDFConfig.check_inputs`
+        config: config (dict): Information necessary for accessing the loaded data. For images this is the file path and
+            for NetCDF files this is a dictionary of configuration values (see
+            :meth:`nccut.netcdfconfig.NetCDFConfig.check_inputs` for structure of dictionary)
+        points: 4 element 1D array with coordinates of the two transect end points: [X1, Y1, X2, Y2]
 
     Returns:
-        2D Dataset from NetCDF file according to selections transposed to (y, x).
+        Subset of data, rescaled points [X1, Y1, X2, Y2], list of [X, Y] rescale factors>. If passed data is an
+        image doesn't do sub-secting since efficiency is less of an issue.
     """
-    # IMPORTANT: Numpy indexes (row, column) with (0, 0) at top left corner
-    if config['z'] == 'N/A':
+    if isinstance(config, str):
+        # Don't subsect if image file
+        img = np.flip(np.asarray(im.open(config)), 0)
+        return img, points, [0, 0]
+    elif config['z'] == 'N/A':
         # 2D NetCDF data
         ds = config['data'][config['var']].rename({config['y']: 'y', config['x']: 'x'})
-        data = ds.sel(y=ds['y'], x=ds['x'])
     else:
         # 3D NetCDF data
         ds = config['data'][config['var']].rename({config['y']: 'y', config['x']: 'x', config['z']: 'z'})
         ds['z'] = ds['z'].astype(str)
-        data = ds.sel(y=ds['y'], x=ds['x'], z=config['z_val'])
-    return data.transpose('y', 'x')
+        ds = ds.sel(z=config["z_val"])
+    ds["x"] = ds["x"].astype(float)
+    ds["y"] = ds["y"].astype(float)
 
+    x_points = np.sort(np.array([points[0], points[2]]))
+    y_points = np.sort(np.array([points[1], points[3]]))
 
-def subset_around_transect(ds, points):
-    """
-    Determines a subset of the data with appropriate margins that surround the transect.
-
-    If room is available around
-
-    Args:
-        ds: Data to be subset from. Either a DataArray (X and Y coordinates must be named "x" and "y") or a 2D array
-        points: 4 element 1D array with coordinates of the two transect end points: [X1, Y1, X2, Y2]
-
-    Returns:
-        Subset of data, rescaled points [X1, Y1, X2, Y2], list of [Y, X] rescale factors
-    """
-    # Determine x, y bounds for subset area around transect
[…]
