+++ tests/test_functions.py
@@ -15,13 +15,13 @@ import xarray as xr
 import nccut.functions as func
 from nccut.multimarker import marker_find
 
-EXAMPLE_JPG_PATH = pooch.retrieve(url="doi:10.5281/zenodo.13883476/example.jpg",
+EXAMPLE_JPG_PATH = pooch.retrieve(url="doi:10.5281/zenodo.14507378/example.jpg",
                                   known_hash="f039e8cb72d6821f4909707767863373230159e384a26ba7edd8a01a3e359e53")
-EXAMPLE_3D_PATH = pooch.retrieve(url="doi:10.5281/zenodo.13883476/example_3d.nc",
-                                 known_hash="a3c946398b1bb5fddf4fbbae36e3fc7e6934cd7b0365d4d9d52b156dadc059d3")
-EXAMPLE_4V_PATH = pooch.retrieve(url="doi:10.5281/zenodo.13883476/example_4v.nc",
+EXAMPLE_3D_PATH = pooch.retrieve(url="doi:10.5281/zenodo.14507378/example_3d.nc",
+                                 known_hash="ccb6c76062d3228799746e68e1bb3ff715538bc3aae796c577c6fb1d06fcdc9f")
+EXAMPLE_4V_PATH = pooch.retrieve(url="doi:10.5281/zenodo.14507378/example_4v.nc",
                                  known_hash="afd261063f4b58c382c46db0d81e69dfb8f5234ef0037b261087177e6d3f7e1b")
-PROJECT_EXAMPLE_PATH = pooch.retrieve(url="doi:10.5281/zenodo.13883476/project_example.json",
+PROJECT_EXAMPLE_PATH = pooch.retrieve(url="doi:10.5281/zenodo.14507378/project_example.json",
                                       known_hash="33c739a4c5515ece45f44eb3de266da46afd6eab47c10cb1f53ac1821226c595")
 
 
@@ -39,8 +39,7 @@ class Test(unittest.TestCase):
 
         # Manual result
         arr = np.asarray(img)
-        rows = np.shape(arr)[0]
-        manual = np.ravel(np.mean(arr[rows - points[3] - 1:rows - points[1], points[0]:points[2]], axis=2))
+        manual = np.ravel(np.mean(arr[points[1]:points[3] + 1, points[0]:points[2]], axis=2))
         # Compare
         self.assertEqual(max(app - manual), 0, "Transect on image not accurate at zero degrees")
 
@@ -57,10 +56,9 @@ class Test(unittest.TestCase):
 
         # Manual result
         arr = np.asarray(img)
-        rows = np.shape(arr)[0]
         ix = np.arange(points[0], points[2])
         iy = np.arange(points[1], points[3])
-        manual = np.ravel(np.mean(arr[rows - iy - 1, ix], axis=1))
+        manual = np.ravel(np.mean(arr[iy, ix], axis=1))
 
         # Compare
         self.assertEqual(max(app - manual), 0, "Transect on image not accurate at 45 degrees")
@@ -78,8 +76,7 @@ class Test(unittest.TestCase):
 
         # Manual result
         arr = np.asarray(img)
-        rows = np.shape(arr)[0]
-        manual = np.ravel(np.mean(np.flip(arr[rows - points[3]:rows - points[1], points[0]:points[2] + 1]), axis=2))
+        manual = np.ravel(np.mean(arr[points[1]:points[3], points[0]:points[2] + 1], axis=2))
 
         # Compare
         self.assertEqual(max(app - manual), 0, "Transect on image not accurate at 90 degrees")
@@ -89,8 +86,8 @@ class Test(unittest.TestCase):
[…]
