### 2024-12-17 7fb7da3
Add json5 to dependencies
