+++ pyproject.toml
@@ -26,6 +26,7 @@ Kivy = "^2.3.0"
 progress = "^1.6"
 plyer = "^2.1.0"
 tomli = "^2.0.2"
+json5 = "^0.10.0"
 pyobjus = {version = "^1.2.3", platform = "darwin"} # Necessary for native file browser on MacOS
 
 [tool.poetry.group.test.dependencies]
