### 2024-12-17 rchartra/NcCut#44 @rchartra
Added json5 to dependency list.
