### 2024-12-18 rchartra/NcCut#43 @platipodium
> will add an error popup so at least the app won't crash and the user can still type in the file path manually.

The popup works for loading a file, but the problem persists on the "save" as well, so for now I cannot save .png / pdf  pictures of the plots....
