### 2024-12-18 d80e73d
Rename transect marker tool to orthogonal chain and transect chain tool to inline chain.
