+++ tests/test_functions.py
@@ -13,16 +13,16 @@ import json
 import pooch
 import xarray as xr
 import nccut.functions as func
-from nccut.multimarker import marker_find
+from nccut.multiorthogonalchain import orthogonal_chain_find
 
-EXAMPLE_JPG_PATH = pooch.retrieve(url="doi:10.5281/zenodo.14507378/example.jpg",
+EXAMPLE_JPG_PATH = pooch.retrieve(url="doi:10.5281/zenodo.14512874/example.jpg",
                                   known_hash="f039e8cb72d6821f4909707767863373230159e384a26ba7edd8a01a3e359e53")
-EXAMPLE_3D_PATH = pooch.retrieve(url="doi:10.5281/zenodo.14507378/example_3d.nc",
+EXAMPLE_3D_PATH = pooch.retrieve(url="doi:10.5281/zenodo.14512874/example_3d.nc",
                                  known_hash="ccb6c76062d3228799746e68e1bb3ff715538bc3aae796c577c6fb1d06fcdc9f")
-EXAMPLE_4V_PATH = pooch.retrieve(url="doi:10.5281/zenodo.14507378/example_4v.nc",
+EXAMPLE_4V_PATH = pooch.retrieve(url="doi:10.5281/zenodo.14512874/example_4v.nc",
                                  known_hash="afd261063f4b58c382c46db0d81e69dfb8f5234ef0037b261087177e6d3f7e1b")
-PROJECT_EXAMPLE_PATH = pooch.retrieve(url="doi:10.5281/zenodo.14507378/project_example.json",
-                                      known_hash="33c739a4c5515ece45f44eb3de266da46afd6eab47c10cb1f53ac1821226c595")
+PROJECT_EXAMPLE_PATH = pooch.retrieve(url="doi:10.5281/zenodo.14512874/project_example.json",
+                                      known_hash='82f37306b94ee54ad1906c6bed72f8c9e8243940f95a8fe1f0d39a27eb920091')
 
 
 class Test(unittest.TestCase):
@@ -152,51 +152,51 @@ class Test(unittest.TestCase):
         self.assertListEqual(list(dat.coords["j"][points[1]:points[3]]), app["j"],
                              "Y Coordinates for NetCDF 90 Degree Transect Incorrect")
 
-    def test_marker_find(self):
+    def test_orthogonal_chain_find(self):
         """
         Test whether valid project files can be accurately identified
         """
         # Data from a valid file is correctly extracted
         proper_json = open(PROJECT_EXAMPLE_PATH)
         proper_data = json.load(proper_json)
-        marker_result = marker_find(proper_data, [], ["Click x", "Click y", "Width"])
-
-        self.assertEqual(len(marker_result), len(proper_data["Vorticity"].keys()), "All markers weren't found")
-        self.assertEqual(len(marker_result[0]), 3, "All fields weren't found")
-        self.assertListEqual(marker_result[0][0], proper_data["Vorticity"]["Marker 1"]["Click x"],
+        chain_result = orthogonal_chain_find(proper_data, [], ["Click x", "Click y", "Width"])
+        self.assertEqual(len(chain_result), len(proper_data["Vorticity"].keys()) - 1, "All chains weren't found")
+        self.assertEqual(len(chain_result[0]), 3, "All fields weren't found")
+        self.assertListEqual(chain_result[0][0], proper_data["Vorticity"]["Orthogonal Chain 1"]["Click x"],
                              "X Coords were not correct")
-        self.assertListEqual(marker_result[1][1], proper_data["Vorticity"]["Marker 2"]["Click y"],
+        self.assertListEqual(chain_result[1][1], proper_data["Vorticity"]["Orthogonal Chain 2"]["Click y"],
[…]
