### 2024-12-18 6d11a86
Image resized to fit screen on window size change.
