### 2024-12-18 rchartra/NcCut#45 @rchartra
Map is now rescaled to fit the viewer window when window is resized.
