### 2024-12-18 540157f
Manual file name entry workaround for filechooser bug
