### 2024-12-18 53277f4
Remove testing statements
