### 2024-12-18 rchartra/NcCut#43 @rchartra
I've added a work around so that if the file browser doesn't work you can still save data/plots by manually entering the file name/path. Still working on getting the file browser to work on all platforms.
