+++ src/nccut/orthogonalchain.py
@@ -7,6 +7,7 @@ Singular orthogonal chain widget.
 
 Graphics and functionality of a singular orthogonal chain created by the orthogonal chain tool.
 """
+import copy
 
 import kivy.uix as ui
 from kivy.metrics import dp
@@ -16,6 +17,23 @@ import math
 import numpy as np
 from kivy.core.window import Window
 import nccut.functions as functions
+
+
+class Click:
+    """
+    Object that mimics a user click.
+
+    Attributes:
+        x (float): X coordinate of click point
+        y (float): Y coordinate of click point
+        pos (tuple): 2 element tuple: (X coord, Y coord)
+    """
+    def __init__(self, x, y):
+        self.x = x
+        self.y = y
+        self.pos = (x, y)
+        self.is_double_tap = False
+        self.button = "left"
 
 
 class OrthogonalChain(ui.widget.Widget):
@@ -32,7 +50,6 @@ class OrthogonalChain(ui.widget.Widget):
         t_width (int): Current width in pixels of orthogonal transects
         home: Reference to root :class:`nccut.homescreen.HomeScreen` instance
         transects (list): List of transects made
-        curr_line: kivy.graphics.Line, Line between cursor and last clicked
         number: kivy.uix.label.Label, Reference to the number label
         size: 2 element array of ints, Size of widget
         pos: 2 element array of ints, Position of widget
@@ -55,7 +72,6 @@ class OrthogonalChain(ui.widget.Widget):
         self.uploaded = False
         self.home = home
         self.transects = []
-        self.curr_line = Line()
         self.number = None
         self.size = self.home.display.size
         self.pos = self.home.display.pos
@@ -93,7 +109,6 @@ class OrthogonalChain(ui.widget.Widget):
         for g in groups:
             for i in g:
                 self.canvas.add(i)
-        self.canvas.add(self.curr_line)
         if self.clicks > 0:
             self.add_widget(self.number)
 
@@ -104,20 +119,15 @@ class OrthogonalChain(ui.widget.Widget):
         Args:
             value (float): New graphics size
         """
-        self.line_width = dp(value / 5)
-        for c in range(1, self.clicks + 1):
-            group = self.canvas.get_group(str(c))
-            for i in group:
-                if isinstance(i, Ellipse):
-                    i.size = (dp(value), dp(value))
-                    i.pos = (i.pos[0] + self.c_size[0] / 2 - dp(value) / 2,
-                             i.pos[1] + self.c_size[1] / 2 - dp(value) / 2)
-                elif isinstance(i, Line):
-                    i.width = self.line_width
-        if self.clicks > 0:
-            self.number.font_size = dp(value) * 2
-        self.curr_line.width = self.line_width
         self.c_size = (dp(value), dp(value))
+        self.line_width = dp(value / 5)
+        points = copy.copy(self.points)
+        for c in range(self.clicks):
+            self.del_point()
+        for p in points:
+            self.update_width(p[2])
+            self.on_touch_down(Click(p[0], p[1]))
+        self.stop_drawing()
 
     def update_width(self, width):
         """
@@ -139,19 +149,20 @@ class OrthogonalChain(ui.widget.Widget):
         """
         self.uploaded = val
 
-    def get_orthogonal(self, line):
+    def get_orthogonal(self, line_start, line_end):
         """
         Get a line orthogonal to line drawn by user to use as transect.
 
         Args:
-            line: kivy.graphics.Line, line between two last clicked points
+            line_start: x, y of start point
+            line_end: x, y of end point
 
         Returns:
             4 element array of floats: Coordinates of the two endpoints of the centered
             orthogonal line with length t_width.
         """
         xyswap = False
-        line = line.points
[…]
