### 2024-12-18 rchartra/NcCut#39 @rchartra
Changed it to dashed lines between clicked points and solid lines for the transects themselves.
