### 2024-12-19 cf791bb
More clear orthogonal transect width selection metrics.
