### 2024-12-19 b63e4a8
Orthogonal transect width popup is readable for all font sizes
