### 2024-12-19 ffad573
Add project uploading functionality to the inline chain tool
