+++ src/nccut/multiorthogonalchain.py
@@ -21,83 +21,6 @@ from nccut.orthogonalchain import OrthogonalChain
 from nccut.orthogonalchainwidth import OrthogonalChainWidth
 
 
-class Click:
-    """
-    Object that mimics a user click.
-
-    Attributes:
-        x (float): X coordinate of click point
-        y (float): Y coordinate of click point
-        pos (tuple): 2 element tuple: (X coord, Y coord)
-    """
-    def __init__(self, x, y):
-        self.x = x
-        self.y = y
-        self.pos = (x, y)
-        self.is_double_tap = False
-        self.button = "left"
-
-
-def correct_test(data, need):
-    """
-    Check if dictionary has necessary fields to be an orthogonal chain
-
-    Args:
-        data (dict): Dictionary to be tested.
-        need (list): List of required fields to qualify as an orthogonal chain:
-            ["Click <cord>", "Click <cord>", "Width"]
-
-    Returns:
-        Boolean, whether dictionary has necessary keys with a list has the value
-    """
-    keys = list(data.keys())
-    if len(keys) == 0:
-        return False
-    else:
-        for item in need:
-            if item not in keys or not isinstance(data[item], list):
-                return False
-    return True
-
-
-def orthogonal_chain_find(data, res, need):
-    """
-    Recursively examines dictionary and finds chain click coordinates and orthogonal transect widths.
-
-    Args:
-        data (dict): Dictionary to examine
-        res (list): Empty list to fill with chain click coordinates and orthogonal transect widths
-        need (list): List of required fields to qualify as an orthogonal chain:
-            ["Click <cord>", "Click <cord>", "Width"]
-
-    Returns:
-        Nested List. A list containing a list for each orthogonal chain which each contains three lists:
-        click X coords, click y coords, and the orthogonal transect width for each click point in the chain.
-        If no qualifying data was found returns empty list. If duplicate data is found (ex: multiple
-        variables in a file) only returns one instance of orthogonal chain data.
-    """
-    for key in list(data.keys()):
-        if key[0:10] == 'Orthogonal':
-            if correct_test(data[key], need):  # Orthogonal chain dict has necessary fields
-                if len(res) == 0:  # If res empty, always add orthogonal chain data
-                    res.append([data[key][need[0]], data[key][need[1]], data[key][need[2]]])
-                else:  # If res not empty, ensure found orthogonal chain data isn't already in res
-                    new = True
-                    for item in res:
-                        l1 = data[key][need[0]]
-                        l2 = item[0]
-                        if len(l1) == len(l2) and len(l1) == sum([1 for i, j in zip(l1, l2) if i == j]):
-                            new = False
-                    if new:
-                        res.append([data[key][need[0]], data[key][need[1]], data[key][need[2]]])
-        else:
-            if type(data[key]) is dict:  # Can still go further in nested dictionary tree
-                orthogonal_chain_find(data[key], res, need)
-            else:
-                return res
-    return res
-
-
 class MultiOrthogonalChain(ui.widget.Widget):
     """
     Orthogonal chain tool widget.
@@ -231,36 +154,15 @@ class MultiOrthogonalChain(ui.widget.Widget):
                 nc_coords = True
             except ValueError:
                 pass
-        found = orthogonal_chain_find(data, [], ["Click " + str(x_name), "Click " + str(y_name), "Width"])
+        found = func.chain_find(data, [], ["Click " + str(x_name), "Click " + str(y_name), "Width"], "Orthogonal")
         if len(found) >= 1:
             if nc_coords:
-                found = self.convert_found_coords(found)
+                found = func.convert_found_coords(found, self.home.display.config)
             self.upload_data(found)
         else:
             content = Label(text="JSON File is not a Project for this File")
[…]
