+++ docs/source/usage.rst
@@ -66,7 +66,7 @@ The :ref:`Inline Chain <inline_chain>` tool allows you to draw multiple transect
 Inline Chain
 ^^^^^^^^^^^^
 
-Using this tool chains of transects can be drawn. Transects will be taken along the line segments drawn between the clicked points. Multiple such *Chains* can be drawn out at once before being plotted.
+Using this tool chains of transects can be drawn. Transects will be taken along the line segments drawn between the clicked points. Multiple such *Chains* can be drawn out at once before being plotted. These *chains* can all be saved together as a *project* and reloaded into the viewer later. When a large file is being explored this allows you to mark out features over multiple sessions.
 
 #. Hit the **Inline Chain** button to enter transect mode.
 #. Click points along the feature you'd like to make transects along. Transects will be made between the points you click.
@@ -75,37 +75,27 @@ Using this tool chains of transects can be drawn. Transects will be taken along 
 #. Right click to select the final point or select **New Chain** to begin a new chain and repeat for as many chains as you'd like.
 #. When done select the **Plot** button and a popup will appear with a plot of all transects from the first chain and downloading options.
 
+#. Click anywhere around the popup or the close button to dismiss
+
 .. _orthogonal_chain:
 
 Orthogonal Chain
 ^^^^^^^^^^^^^^^^
 
-Using this tool multiple *chains* can be drawn onto a loaded file. These *chains* can all be saved together as a *project* and reloaded into the viewer later. When a large file is being explored this allows you to mark out features over multiple sessions.
+Using this tool transects will be made orthogonally to the line segments in the chain marked out by the user with a specified width. Multiple *chains* can be drawn onto a loaded file at once before being plotted. These *chains* can all be saved together as a *project* and reloaded into the viewer later. When a large file is being explored this allows you to mark out features over multiple sessions.
 
 #. Hit the **Orthogonal Chain** button to enter transect mode.
-#. Click points along the feature you'd like to make transects across. Dots will appear on either side of the line drawn indicating the start and end points of the transects that will be made.
+#. Click points along the feature you'd like to make transects across. Solid lines will appear crossing the dashed line you draw indicating where the transect will be taken.
 
     * Select **Drag Mode** from the actions sidebar to drag the image without selecting points, and select **Transect Mode** to go back.
     * Select **Edit Mode** to delete either the last point clicked or the last chain drawn. You can also press the esc key to delete the last chain drawn.
     * To change the width of the orthogonal transects made you may enter the number of pixels into the **Width** text box. Select **Set** to change the width for all future transects.
 
         * This will not change the width of transects already drawn
-        * The default width is 40 pixels, you can enter any width within 0 and 400
         * If you change the width of a chain any new chains will continue to use that same width unless you change it again.
 
 #. Right click to select the final point or select **New Chain** to begin a new chain and repeat for as many chains as you'd like.
 #. When done select the **Plot** button and a popup will appear with a plot of all transects from the first chain and downloading options.
-
-#. Click anywhere around the popup or the close button to dismiss
-
-To upload a previously worked on project:
-
-#. Load the same dataset/image you worked on previously. Project files are specific to the file, variable, and coordinate
-selections originally used.
-#. Hit the **Orthogonal Chain** button to enter transect mode.
-#. Instead of clicking new points, select the **Upload Project** button.
-#. Enter the file name of the transect data you saved previously and select **Ok**.
-#. All chains from the file will load onto the viewer and you can continue working on the project.
 
 Plotting
 --------
@@ -128,6 +118,18 @@ Plotting
     * **Save All Z Values** will save all z values for the selected transects and variables. This option only appears if using a NetCDF file with three dimensions.
[…]
