+++ docs/source/example.rst
@@ -1,11 +1,12 @@
 Examples
 =========
 
-To test out NcCut you can download some example files from here: `https://zenodo.org/records/13883476 <https://zenodo.org/records/13883476>`_:
+To test out NcCut you can download some example files from here: `https://zenodo.org/records/14525966 <https://zenodo.org/records/14525966>`_:
 
 
 * ``example.jpg`` is a SAR image of sea ice floes in the arctic ocean from the Sentinel-1 satellite
-* ``project_example.json`` is an example project file output from ``example_4v.nc`` to practice uploading projects
+* ``orthogonal_project_example.json`` is an example project file output from ``example_4v.nc`` to practice uploading projects with the orthogonal chain tool.
+* ``inline_project_example.json`` is an example project file output from ``example_4v.nc`` to practice uploading projects with the inline chain tool.
 * ``example_3d.nc`` is a subsection of a modeled ocean temperature dataset from the MITgcm LLC4320 Model near the gulf stream
 * ``example_4v.nc`` is a dataset of various ocean circulation metrics calculated from velocity data from the MITgcm LLC4320 Model near the gulf stream
 * ``test_img.png`` is simply an image of a square to use to validate the transect tools' accuracy
@@ -16,7 +17,7 @@ To test out NcCut you can download some example files from here: `https://zenodo
 Working with the Output Data
 ----------------------------
 
-This code example uses ``example_4v.nc`` and ``project_example.json``
+This code example uses ``example_4v.nc`` and ``orthogonal_project_example.json``
 
 
 .. toctree::
