### 2024-12-20 rchartra/NcCut#41 @rchartra
I've made the width adjustment more clear as to the units and it now alerts the user of incorrect input. Still need to add axes ticks/labels.
