### 2024-12-23 9d9e257
X and Y axis ticks
