### 2024-12-28 rchartra/NcCut#41 @rchartra
Axis ticks and labels added.
