### 2025-01-19 f218c16
Fix mac bugs
