+++ src/nccut/multiinlinechain.py
@@ -107,7 +107,7 @@ class MultiInlineChain(ui.widget.Widget):
             if platform.system() == "Darwin":
                 # Construct the AppleScript command for selecting json files
                 script = """
-                        set file_path to choose file of type {"json"}
+                        set file_path to choose file of type {"public.json"}
                         POSIX path of file_path
                         """
                 result = subprocess.run(
@@ -117,9 +117,6 @@ class MultiInlineChain(ui.widget.Widget):
                 if result.returncode == 0:
                     file_path = result.stdout.strip()
                     self.check_file(file_path)
-                else:
-                    func.alert_popup("Cannot upload project files at this time. Error: "
-                                     + str(result.stderr))
             else:
                 path = filechooser.open_file(filters=["*.json"])
                 if path is not None and len(path) != 0:
