### 2025-01-22 6e9c0fd
Fix image not resizing fully on start bug
