### 2025-01-23 rchartra/NcCut#43 @rchartra
I've switched to a method that does not use any non-standard Python packages. I've tested it on several Macs and have seen no issues with it so am closing this issue. Please reopen if you continue to run into issues.
