### 2025-01-24 rchartra/NcCut#16 @platipodium
any updates on this?  For me, the UI has gotten much more usable already, thanks for the effort!
