### 2025-01-24 #7185 @platipodium
- @hugoledoux This concludes my review.  I have completed my checklist and I recommend publishing in JOSS.
- @rchartra Please fix formatting of your first reference for the QGIS Contributors
