### 2025-01-24 rchartra/NcCut#16 @rchartra
That's great to hear! I'm meeting with a few people next week to try out the tool with the new updates. Once I get their feedback I'll implement any additional changes and close this issue.
