+++ paper.bib
@@ -32,7 +32,7 @@
 }
 
 @software{Panoply,
-  author = {NASA Goddard Institute for Space Studies},
+  author = {{NASA Goddard Institute for Space Studies}},
   title = {Panoply Data Viewer},
   year = {2023},
   date = {2023-09-22},
@@ -82,7 +82,7 @@
 }
 
 @software{qgis,
-  author       = {QGIS Contributors},
+  author       = {{QGIS Contributors}},
   title        = {QGIS},
   month        = jan,
   year         = 2022,
