### 2025-02-03 rchartra/NcCut#52 @Huite
Docs mention `~/.config/your_app/config.toml`
Related to: https://github.com/openjournals/joss-reviews/issues/7185

https://nccut.readthedocs.io/en/latest/usage.html#configuration-file

Point 4 states; Either ‘%APPDATA%nccutnccut_config.toml’ on Windows or ‘~/.config/your_app/config.toml’ on Linux and macOS. I reckon "your_app" here is copied from a template, since the source shows: `config_path = Path.home() / ".config" / "nccut" / "nccut_config.toml"`
