### 2025-02-03 rchartra/NcCut#53 @Huite
"Export" then "Upload Project" is confusing to me
Related to: https://github.com/openjournals/joss-reviews/issues/7185

I think it's the "upload" word that has me a bit confused: to me it wasn't clear whether it was opening a file, or saving one. I tried "storing" a new file repeatedly, and it'll give the error "Cannot upload project files at this time".

The documentation isn't quite clear, and it has some markup issues too:

```
#. Load the same dataset/image you worked on previously. Project files are specific to the file, variable, and coordinate selections originally used. #. Hit the tool button of the tool you used previously to enter transect mode. #. Instead of clicking new points, select the Upload Project button. #. Enter the file name of the transect data you saved previously from the plotting menu and select Ok. #. All chains from the file will load onto the viewer and you can continue working on the project.
```

The button is positioned in the general screen; the saving is found "behind" the plot button.

I think I understand why the export is "behind" the plot button, since the plot menu controls the variables?
My problem is that I didn't make the connection between "Export Data" and "Upload Project".

Ideally, you could move the opening and saving buttons to the same menu. But at the least the terms should be the same (i.e. data versus project). Second I think I'd prefer "open" instead of "upload" (upload has clear "uploading to the internet" associations for me).
