### 2025-02-03 #7185 @Huite
Apologies for the delay! (I wanted to look start of January, but my wife and I have been enjoying the company of a baby girl since the end of October and I got my year off to a slightly messy start.)

I've just reported two minor remaining issues, one docs, one usability related.

I have completed my checklist and I recommend publishing in JOSS.
