### 2025-02-04 d7a64cc
Docs config file location typo
