### 2025-02-04 rchartra/NcCut#53 @rchartra
These are good points and I agree "Upload Project" is a bit unclear and out of place with the rest of the app's nomenclature. I also think it's reasonable to be able to save chain data without first plotting it. Here's my proposed solution:

- Leave export data buttons in plotting menu as they are
- Beneath the "Tool Actions" section of the tool sidebar have a "Chain Data" section containing the buttons:
    - Open Data: Load previous chains from previously exported data
    - Plot Data: Open plotting menu and more selective/inclusive export options
    - Export Data: Save all current chains for the currently loaded variable and z level
