+++ paper.md
@@ -33,7 +33,7 @@ experience to use.  In addition to being a GUI for generating "quick" visualizat
 user to view and produce plots of transects of NetCDF4 data along various dimensions.  This is a 
 commonly performed task in oceanography and atmospheric sciences research used to understand the vertical composition of 
 a section of ocean or atmosphere. For example, @Sivkov:2021 used such vertical transects to study the vertical 
-distributions of suspended particulate matter across the Atlantic ocean. In another example, @Gutjahr:2022 used vertical 
+distributions of suspended particulate matter across the Atlantic Ocean. In another example, @Gutjahr:2022 used vertical 
 transects of modeled climate and ocean data in Greenland to study air-sea interactions during a Katabatic storm.
 
 Such transect selecting tools are a common feature in GIS software (@qgis, @transectizer, @GRASS_GIS) and are present in 
@@ -50,27 +50,27 @@ NcCut’s functions can be divided into two main categories:
 
 ## Viewer
 
-NcCut allows users to load NetCDF4 files, or images in .JPG, .JPEG, or .PNG formats, and has adjustable graphics for 
+NcCut allows users to load NetCDF4 files, or images in JPG, JPEG, or PNG formats, and has adjustable graphics for 
 different file sizes. In the case of NetCDF4 files, the user can choose between different variables and dimensions of 
 the data to view. The user can select which dimensions to use as the x, y and z dimensions for the viewer, and once 
-displayed the user can switch between different displays of the x, y dimensions at different z values. This is intuitive 
+displayed the user can switch between different displays of the x-y dimensions at different z-values. This is intuitive 
 for many NetCDF4 files which contain three-dimensional datasets with data arrays at different vertical heights or 
 depths. The data is displayed as an image mapped to a color map, which can be changed by the user as well as the 
-contrast of the image. Once loaded and configured users can drag, rotate, flip, and zoom in on the file however they 
+contrast of the image. Once loaded and configured, users can drag, rotate, flip, and zoom in on the file however they 
 please. The NetCDF4 viewing capabilities and settings are shown in Figure 1, where an ocean temperature dataset from the MIT General Circulation 
-Model LLC4320 simulation is being displayed (accessed via \verb|xmitgcm| https://xmitgcm.readthedocs.io/en/latest/index.html).
+Model LLC4320 simulation is being displayed (accessed via \verb|xmitgcm| https://xmitgcm.readthedocs.io).
 
-![a.) The viewing interface shown with a NetCDF4 file of ocean temperature in the gulf stream region. b.) The setting options for NetCDF files. c.) The initial configuration 
+![a.) The viewing interface shown with a NetCDF4 file of ocean temperature in the Gulf Stream region. b.) The setting options for NetCDF files. c.) The initial configuration 
 menu for loading NetCDF4 files. ](images/paper_fig_1.png)
 
 ## "Cut" Making
 
-Using NcCut, users can mark out transects onto loaded images or NetCDF4 data which can then be plotted and saved for further analysis. There two tools for marking out transects:
+Using NcCut, users can mark out transects onto loaded images or NetCDF4 data which can then be plotted and saved for further analysis. There are two tools for marking out transects:
 
-The first is the 'Inline Chain' tool. Users can click on points along a features and lines will be drawn connecting the points forming a "chain". Transects are then made along the line segments shown (Figure 2a). Multiple chains can be drawn at once, and plotted together in the plotting menu (Figure 3). When plotted, the transect data from each segment is plotted continuously end to end. If the loaded file is an image, the transect data is the mean of the RGB values of each pixel along the transect. The pixels are 
[…]
