### 2025-02-12 f78183f
Change "upload" to "reload" and "project" to "chain data". Data loading bug fixes. Better organization of tool buttons.
