+++ src/nccut/filedisplay.py
@@ -13,6 +13,9 @@ ScatterLayout class. Manages the creation and deletion of tools.
 import kivy
 from kivy.graphics import Color, Line
 from kivy.graphics.transformation import Matrix
+from kivy.uix.button import Button
+from kivy.uix.popup import Popup
+from kivy.uix.textinput import TextInput
 from kivy.uix.scatterlayout import ScatterLayout
 from kivy.core.window import Window
 import kivy.uix as ui
@@ -21,10 +24,17 @@ from kivy.core.image import Image as CoreImage
 from kivy.metrics import dp
 from PIL import Image as im
 from PIL import ImageEnhance
+import platform
+import subprocess
+from plyer import filechooser
 import numpy as np
 from scipy.interpolate import RegularGridInterpolator
 import matplotlib.pyplot as plt
 import io
+import os
+import pathlib
+import re
+import copy
 import warnings
 import nccut.functions as func
 from nccut.multiorthogonalchain import MultiOrthogonalChain
@@ -44,6 +54,7 @@ class FileDisplay(ScatterLayout):
             value is the necessary configuration settings. For images, the config dictionary has form
             {"image": str(file_path)}. For a netcdf file the value is a dictionary of configuration values (see
             :meth:`nccut.netcdfconfig.NetCDFConfig.check_inputs` for structure of dictionary)
+        default_orthogonal_width: Default width for the orthogonal transect tool
         f_type (str): File type being loaded ("image" or "netcdf")
         home: Reference to root :class:`nccut.homescreen.HomeScreen` instance
         pos (tuple): Position of viewer. Used to properly place transect widgets on screen.
@@ -70,15 +81,23 @@ class FileDisplay(ScatterLayout):
         cmaps (dict): Dictionary of colormap names mapping to colormap values from cv2
         colormap: current colormap data
         btn_height: Height for buttons in sidebar which adapts to font size
+        tools_lbl: "Tools" sidebar label
         inline_chain_btn: Inline chain button which opens inline chain tool
-        orthogonal_chain__btn: Orthogonal chain button which opens orthogonal chain tool
-        initial_side_bar (list): List of sidebar buttons in 'Tools' sidebar menu
+        orthogonal_chain_btn: Orthogonal chain button which opens orthogonal chain tool
+        initial_side_bar (list): List of sidebar buttons in 'Transect Tools' sidebar menu
+        tool_actions_lbl: "Tool Actions" sidebar label
+        new_chain_btn: New Chain button to be loaded when in transect mode
         drag_btn: Drag button to be loaded when in transect mode
         edit_btn: Edit button to be loaded when in transect mode
         close_tool_btn: Button to close currently loaded tool
-        action_widgets: Initial list of widgets in Actions menu, not including any widgets added by the tool
-        tool_action_widgets: List of widgets in Actions menu at any given moment
+        chain_data_lbl: "Chain Data" sidebar label
+        open_data_btn: Button to open previously exported chain data
+        tool_sb_widgets_constant: Initial list of widgets in main tool meny menu, not including any widgets added by the
+            tool
+        tool_sb_widgets: List of widgets in main tool menu at any given moment
+        drag_mode_lbl: "Drag Mode" sidebar label
         tran_mode_btn: Button to close Drag Mode
+        edit_mode_lbl: "Edit Mode" sidebar label
         back_btn: Back button for editing mode
         delete_chain_btn: Delete chain button
         delete_point_btn: Delete point button
@@ -134,6 +153,7 @@ class FileDisplay(ScatterLayout):
         self.colormap = g_config["colormap"]
         self.btn_height = dp(20) + self.home.font
         # Initial Sidebar Widgets
[…]
