+++ tests/test_gui.py
@@ -32,10 +32,10 @@ EXAMPLE_3D_PATH = pooch.retrieve(url="doi:10.5281/zenodo.14525966/example_3d.nc"
                                  known_hash="ccb6c76062d3228799746e68e1bb3ff715538bc3aae796c577c6fb1d06fcdc9f")
 EXAMPLE_4V_PATH = pooch.retrieve(url="doi:10.5281/zenodo.14525966/example_4v.nc",
                                  known_hash="afd261063f4b58c382c46db0d81e69dfb8f5234ef0037b261087177e6d3f7e1b")
-ORTHOGONAL_PROJECT_EXAMPLE_PATH = pooch.retrieve(url="doi:10.5281/zenodo.14525966/orthogonal_project_example.json",
-                                                 known_hash='82f37306b94ee54ad1906c6bed72f8c9e8243940f95a8fe1f0d39a27eb920091')
-INLINE_PROJECT_EXAMPLE_PATH = pooch.retrieve(url="doi:10.5281/zenodo.14525966/inline_project_example.json",
-                                             known_hash='9f0f1d4d536cc445ccfc5aa07a011a32379673a114fbe2b7be7070bd2f73e9b5')
+ORTHOGONAL_CHAIN_DATA_EXAMPLE_PATH = pooch.retrieve(url="doi:10.5281/zenodo.14525966/orthogonal_project_example.json",
+                                                    known_hash='82f37306b94ee54ad1906c6bed72f8c9e8243940f95a8fe1f0d39a27eb920091')
+INLINE_CHAIN_DATA_EXAMPLE_PATH = pooch.retrieve(url="doi:10.5281/zenodo.14525966/inline_project_example.json",
+                                                known_hash='9f0f1d4d536cc445ccfc5aa07a011a32379673a114fbe2b7be7070bd2f73e9b5')
 
 
 class AppInfo:
@@ -96,8 +96,11 @@ def load_3d_nc(z_val):
 def select_sidebar_button(text):
     sidebar = run_app.home.ids.dynamic_sidebar
     but = next((but for but in sidebar.children if isinstance(but, Button) and but.text == text), None)
-    but.dispatch('on_press')
-    but.dispatch('on_release')
+    if but:
+        but.dispatch('on_press')
+        but.dispatch('on_release')
+    else:
+        raise Exception("Button not in Sidebar")
 
 
 class Test(unittest.TestCase):
@@ -118,7 +121,7 @@ class Test(unittest.TestCase):
         run_app.home.ids.file_in.text = ""
         run_app.home.load_btn()
         self.assertEqual(run_app.home.children[0].text, "Invalid File Name", "Empty file names are invalid")
-        run_app.home.ids.file_in.text = ORTHOGONAL_PROJECT_EXAMPLE_PATH
+        run_app.home.ids.file_in.text = ORTHOGONAL_CHAIN_DATA_EXAMPLE_PATH
         run_app.home.load_btn()
         self.assertEqual(run_app.home.children[0].text, "Unsupported File Type", "No unaccepted file types")
         run_app.home.ids.file_in.text = "teacup.jpg"
@@ -191,11 +194,11 @@ class Test(unittest.TestCase):
         self.assertIsNone(run_app.home.settings_bar.netcdf_btn.parent, "NetCDF settings menu was not removed")
         self.assertEqual(len(run_app.home.display.children), 1, "Not all tools were removed from display")
 
-    def test_project_upload(self):
+    def test_data_load(self):
         """
-        Tests the project upload function of the tools.
+        Tests the chain data load function of the tools.
 
-        Assumes project files are valid project files for 'support/example_4v.nc' with
+        Assumes chain data files are valid for 'support/example_4v.nc' with
[…]
