+++ docs/source/example.rst
@@ -5,12 +5,12 @@ To test out NcCut you can download some example files from here: `https://zenodo
 
 
 * ``example.jpg`` is a SAR image of sea ice floes in the arctic ocean from the Sentinel-1 satellite
-* ``orthogonal_project_example.json`` is an example project file output from ``example_4v.nc`` to practice uploading projects with the orthogonal chain tool.
-* ``inline_project_example.json`` is an example project file output from ``example_4v.nc`` to practice uploading projects with the inline chain tool.
+* ``orthogonal_project_example.json`` is an example chain data file output from ``example_4v.nc`` to practice loading chain data with the orthogonal chain tool.
+* ``inline_project_example.json`` is an example chain data file output from ``example_4v.nc`` to practice loading chain data with the inline chain tool.
 * ``example_3d.nc`` is a subsection of a modeled ocean temperature dataset from the MITgcm LLC4320 Model near the gulf stream
 * ``example_4v.nc`` is a dataset of various ocean circulation metrics calculated from velocity data from the MITgcm LLC4320 Model near the gulf stream
 * ``test_img.png`` is simply an image of a square to use to validate the transect tools' accuracy
-* ``test_project_file.json`` is another example project file for practicing uploading project files
+* ``test_project_file.json`` is another example chain data file for practicing uploading chain data files
 
 .. _output:
 
