### 2025-02-13 eaafb50
Minor doc and paper typos
