### 2025-02-13 rchartra/NcCut#53 @rchartra
I've implemented all of the above except for adding a button for exporting data from the tool actions side bar. I'm still working on that, it's just going to take a minute so I just wanted to give an update until then. I also fixed the docs markdown issue.
