### 2025-02-13 #7185 @rchartra
I've completed the above:

- Version tag: v1.0.0
- Zenodo DOI: 10.5281/zenodo.14862691

Thank you @platipodium and @Huite for all your suggestions, it's been great working with you both!
