### 2025-02-13 #7185 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
