### 2025-02-19 #7185 @kthyng
@rchartra Looks like a CC license has been added to the repository — Can you update me on the reasoning? CC is not an open source license.
