### 2025-02-19 #7185 @kthyng
Please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file.
