### 2025-02-20 #7185 @rchartra
The CC0 1.0 license was recommended to me as being better suited for images. As such per my REUSE.toml configuration only those files are covered by the CC0 1.0 license while all code is covered by the BSD-3 license
