### 2025-02-20 63f1c97
Fix capitalization in bib file
