+++ paper.bib
@@ -3,7 +3,7 @@
   year = {2022},
   month = {04},
   pages = {e2021JC018075},
-  title = {Air‐Sea Interactions and Water Mass Transformation During a Katabatic Storm in the Irminger Sea},
+  title = {{Air‐Sea Interactions and Water Mass Transformation During a Katabatic Storm in the Irminger Sea}},
   volume = {127},
   journal = {Journal of Geophysical Research: Oceans},
   doi = {10.1029/2021JC018075}
@@ -11,7 +11,7 @@
 
 @article{Sivkov:2021,
   author = {Sivkov, V. and Bubnova, E.},
-  title = {Distribution of suspended particulate matter at the equatorial transect in the Atlantic Ocean},
+  title = {{Distribution of Suspended Particulate Matter at the Equatorial Transect in the Atlantic Ocean}},
   journal = {Ocean Science},
   volume = {17},
   number = {5},
@@ -22,7 +22,7 @@
 
 @software{ncvue,
   author = {Matthias Cuntz},
-  title = {{ncvue: a minimal GUI for a quick view of netcdf files}},
+  title = {{ncvue: A Minimal GUI for a Quick View of NetCDF Files}},
   month = oct,
   year = 2021,
   publisher = {Zenodo},
@@ -33,7 +33,7 @@
 
 @software{Panoply,
   author = {{NASA Goddard Institute for Space Studies}},
-  title = {Panoply Data Viewer},
+  title = {{Panoply Data Viewer}},
   year = {2023},
   date = {2023-09-22},
   version = {5.2.9},
@@ -42,7 +42,7 @@
 
 @software{Ncview,
   author = {David W. Pierce},
-  title = {Ncview},
+  title = {{Ncview}},
   year = {2021},
   date = {2023-09-22},
   version = {2.1.9},
@@ -51,7 +51,7 @@
 
 @software{odv,
   author       = {Schlitzer, Reiner},
-  title        = {Ocean Data View, Version 5.6.2},
+  title        = {{Ocean Data View, Version 5.6.2}},
   month        = may,
   year         = 2023,
   publisher    = {Zenodo},
@@ -60,7 +60,7 @@
 }
 
 @Manual{GRASS_GIS,
-  title = {Geographic Resources Analysis Support System (GRASS GIS) Software, Version 8.2},
+  title = {{Geographic Resources Analysis Support System (GRASS GIS) Software, Version 8.2}},
   author = {{GRASS Development Team}},
   organization = {Open Source Geospatial Foundation},
   address = {USA},
@@ -71,8 +71,8 @@
 
 @software{transectizer,
   author       = {Jorge Tornero},
-  title        = {{Transectizer, a QGIS plugin for linear transect
-                   design}},
+  title        = {{Transectizer, a QGIS Plugin for Linear Transect
+                   Design}},
   month        = may,
   year         = 2014,
   publisher    = {Zenodo},
