### 2025-02-20 #7185 @rchartra
> Please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file.

Fixed!
