+++ src/nccut/multiinlinechain.py
@@ -24,7 +24,8 @@ class MultiInlineChain(ui.widget.Widget):
         c_on (bool): Whether there are any chains active
         load_fail (bool): If anything has gone wrong in the chain data loading process
         home: Reference to root :class:`nccut.homescreen.HomeScreen` instance
-        d_btn: RoundedButton, Plot button to activate PlotPopup
+        p_btn: RoundedButton, Plot button to activate PlotPopup
+        e_btn: RoundedButton, Button to export all chain data to JSON
         dragging (bool): Whether viewer is in dragging mode
         clicks (int): Number of clicks made by user. Does not decrease when points are deleted
             unless all points are deleted in which case it goes back to zero.
@@ -41,8 +42,10 @@ class MultiInlineChain(ui.widget.Widget):
         self.c_on = False
         self.load_fail = False
         self.home = home
-        self.d_btn = func.RoundedButton(text="Plot Data", size_hint_y=None, height=b_height, font_size=self.home.font)
-        self.d_btn.bind(on_press=lambda x: self.gather_popup())
+        self.p_btn = func.RoundedButton(text="Plot Data", size_hint_y=None, height=b_height, font_size=self.home.font)
+        self.p_btn.bind(on_press=lambda x: self.gather_popup())
+        self.e_btn = func.RoundedButton(text="Export Data", size_hint_y=None, height=b_height, font_size=self.home.font)
+        self.e_btn.bind(on_press=lambda x: func.ask_for_output_file_name("*.json", self.save_data, self.home))
         self.dragging = False
         self.clicks = 0
 
@@ -52,7 +55,8 @@ class MultiInlineChain(ui.widget.Widget):
         Args:
             font (float): New font size
         """
-        self.d_btn.font_size = font
+        self.p_btn.font_size = font
+        self.e_btn.font_size = font
 
     def update_l_col(self, color):
         """
@@ -149,8 +153,11 @@ class MultiInlineChain(ui.widget.Widget):
                     chain.on_touch_down(touch)
                     self.clicks += 1
                 chain.load_mode(False)
-                if self.clicks >= 2 and self.d_btn.parent is None:
-                    self.home.display.add_to_sidebar(self.d_btn)
+                if self.clicks >= 2:
+                    if self.p_btn.parent is None:
+                        self.home.display.add_to_sidebar(self.p_btn)
+                    if self.e_btn.parent is None:
+                        self.home.display.add_to_sidebar(self.e_btn)
                 if self.load_fail:  # If load goes wrong, stop and undo everything
                     self.undo_load(c)
                     return
@@ -170,8 +177,10 @@ class MultiInlineChain(ui.widget.Widget):
         if len(self.children) == 0:
             # Remove sidebar buttons if deleted chain was the only chain
             self.clicks = 0
-            if self.d_btn in self.home.display.tool_sb_widgets:
-                self.home.display.remove_from_tool_sb_widgets(self.d_btn)
+            if self.p_btn in self.home.display.tool_sb_widgets:
+                self.home.display.remove_from_tool_sb_widgets(self.p_btn)
+            if self.e_btn in self.home.display.tool_sb_widgets:
+                self.home.display.remove_from_tool_sb_widgets(self.e_btn)
             if self.dragging:
                 self.home.display.drag_mode()
             self.new_chain()
@@ -195,8 +204,10 @@ class MultiInlineChain(ui.widget.Widget):
         if len(self.children) == 0:
             # Remove sidebar buttons if deleted inline chain was the only inline chain
             self.clicks = 0
-            if self.d_btn in self.home.display.tool_sb_widgets:
-                self.home.display.remove_from_tool_sb_widgets(self.d_btn)
+            if self.p_btn in self.home.display.tool_sb_widgets:
[…]
