+++ src/nccut/plotpopup.py
@@ -16,8 +16,6 @@ from kivy.uix.popup import Popup
 from kivy.uix.label import Label
 from kivy.uix.checkbox import CheckBox
 from kivy.uix.dropdown import DropDown
-from kivy.uix.textinput import TextInput
-from plyer import filechooser
 import nccut.functions as func
 from nccut.plotwindow import PlotWindow
 from kivy.core.image import Image as CoreImage
@@ -27,13 +25,8 @@ from scipy.interpolate import RegularGridInterpolator
 import numpy as np
 import copy
 import pandas as pd
-import datetime
 import json
 import pathlib
-import os
-import re
-import platform
-import subprocess
 
 
 KV_FILE_PATH = pathlib.Path(__file__).parent.resolve() / "plotpopup.kv"
@@ -424,7 +417,7 @@ class PlotPopup(Popup):
                     final[var] = {}
                     for z in list(dat[var].keys()):
                         final[var][z] = self.add_group_info(dat[var][z])
-            final = self.add_metadata(final)
+            final = func.add_metadata(self.config, self.f_type, self.home, final)
             with open(f_path, "w") as f:
                 json.dump(final, f)
 
@@ -451,42 +444,6 @@ class PlotPopup(Popup):
         except Exception as error:
             func.alert_popup(str(error))
 
-    def add_metadata(self, dicti):
-        """
-        Adds global and variable specific data to an output dictionary.
-
-        Args:
-            dicti: Dictionary of data about to be exported
-
-        Returns:
-            dicti: Dictionary of data with metadata fields added
-        """
-        def attrs_to_str(d):
-            return {k: str(v) for k, v in d.items()}
-
-        config = self.config[self.f_type]
-        # On GitHub Linux Runner a user is not defined resulting in an error
-        try:
-            user = os.getlogin()
-        except OSError:
-            user = "_user_id_not_found_"
-        global_metadata = {"time_stamp": datetime.datetime.now().strftime("%m/%d/%Y, %H:%M:%S"),
-                           "user": user, "license": "CC0-1.0"}
-        global_metadata.update(self.home.general_config["metadata"])
-        if self.f_type == "netcdf":
-            global_metadata["file"] = config["file"]
-            global_metadata["netcdf_attrs"] = attrs_to_str(config["data"].attrs)
-            dims = {config["x"]: attrs_to_str(config["data"][config["x"]].attrs),
-                    config["y"]: attrs_to_str(config["data"][config["y"]].attrs)}
-            if config["z"] != "N/A":
-                dims[config["z"]] = attrs_to_str(config["data"][config["z"]].attrs)
-            global_metadata["dim_attrs"] = dims
-            for key in list(dicti.keys()):
-                dicti[key][key + "_attrs"] = attrs_to_str(config["data"][key].attrs)
-        else:
-            global_metadata["file"] = config
-        dicti["global_metadata"] = global_metadata
-        return dicti
 
     def add_group_info(self, dicti):
         """

[…]
